% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateDistances.R
\name{calculateDistances}
\alias{calculateDistances}
\title{Estimate minimum traveled distances}
\usage{
calculateDistances(
  data,
  land.shape = NULL,
  epsg.code = NULL,
  grid.resolution = 100,
  mov.directions = 16,
  id.col = getDefaults("id"),
  lon.col = getDefaults("lon"),
  lat.col = getDefaults("lat"),
  verbose = TRUE
)
}
\arguments{
\item{data}{A data frame with animal positions, containing longitude and latitude values (unprojected).}

\item{land.shape}{A  shape file containing coastlines.}

\item{epsg.code}{Coordinate reference system used to project positions (class 'CRS').
If not supplied, CRS is assumed to be the same as in land.shape.}

\item{grid.resolution}{Grid cell size in meters  over which shortest paths are going to be estimated.}

\item{mov.directions}{Number of directions allowed for shortest path calculation.
Passed to \link[gdistance]{transition}.}

\item{id.col}{Name of the column containing animal IDs. Defaults to 'ID'.}

\item{lon.col}{Name of the column containing longitude values (unprojected). Defaults to 'lon'.}

\item{lat.col}{Name of the column containing latitude values (unprojected). Defaults to 'lat'.}

\item{verbose}{Output process info and progress bar to console? Defaults to TRUE.}
}
\value{
A data frame containing (minimum) distances traveled between each consecutive animal position,
calculated separately for each individual.
}
\description{
Function to calculate the shortest path in water between consecutive positions.
If no land.shape is provided, linear ('great-circle') distances are calculated instead.
}
