% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateFrequencies.R
\name{calculateFrequencies}
\alias{calculateFrequencies}
\title{Calculate hourly detection/presence frequencies}
\usage{
calculateFrequencies(
  table,
  n.individuals,
  tagging.dates,
  type = "detections",
  subset = NULL
)
}
\arguments{
\item{table}{A data frame object containing binned detections in the wide format
(time bin x individual matrix, with values corresponding to the number of detections),
as returned by \code{\link{createWideTable}} with value.type="detections".}

\item{n.individuals}{Number of tagged individuals.}

\item{tagging.dates}{A POSIXct vector containing the tag/release date of each animal.}

\item{type}{Either "detections" or "presences".}

\item{subset}{If defined, hourly frequencies are calculated independently for each level
of this variable. If left NULL, single detection/presence frequencies are calculated for the
whole monitoring period.}
}
\value{
Hourly frequences for each individual.
}
\description{
Estimates hourly detection frequencies / presence probabilities of each
individual. If a subset variable is defined, frequencies are calculated
independently by group.
}
