% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateKUDs.R
\name{calculateKUDs}
\alias{calculateKUDs}
\title{Calculate animals' kernel utilization areas}
\usage{
calculateKUDs(
  data,
  subset = NULL,
  bandwidth,
  grid,
  land.shape,
  id.col = "ID",
  lon.col = "lon",
  lat.col = "lat"
)
}
\arguments{
\item{data}{A data frame with animal positions (COAs), containing 'longitude' and 'latitude' columns.}

\item{subset}{If defined, KUDs are calculated independently for each level of this variable.
If left NULL, single KUDs are calculated for the whole monitoring period.}

\item{bandwidth}{Kernel bandwidth (h) - smoothing parameter used to calculate KUDs.}

\item{grid}{SpatialPixels object containing the grid over which animal KUDs are going to be estimated
(see grid argument in the \code{\link[adehabitatHR]{kernelUD}} function).}

\item{land.shape}{A shape file containing coastlines.}

\item{id.col}{Name of the column containing animal IDs. Defaults to "ID".}

\item{lon.col}{Name of the column containing projected longitudes. Defaults to "lon".}

\item{lat.col}{Name of the column containing projected latitudes. Defaults to "lat".}
}
\value{
A list containing the estimated UDs, areas corresponding to
50\% and 95\% of occurrence probability, and a formatted data frame.
}
\description{
Pipeline to calculate animals' utilization areas using kernel
density estimation, based on \code{\link[adehabitatHR]{kernelUD}} function. It estimates
areas corresponding to 50\% and 95\% of occurrence probability (automatically subtracting
portions over land masses) and allows estimation of KUDs independently by group.
}
\seealso{
\code{\link[adehabitatHR]{kernelUD}}
}
