% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateOverlap.R
\name{calculateOverlap}
\alias{calculateOverlap}
\title{Calculate pairwise spatio-temporal overlaps (association index)}
\usage{
calculateOverlap(
  table,
  id.metadata,
  id.groups = NULL,
  subset = NULL,
  id.col = "ID",
  tagdate.col = "tagging_date",
  group.comparisons = "all"
)
}
\arguments{
\item{table}{A data frame object containing binned detections in the wide format
(time bin x individual matrix, with values corresponding to the receiver with the
highest number of detections), as returned by \code{\link{createWideTable}}.}

\item{id.metadata}{A data frame containing individual's metadata. It should contain
at least  animal IDs  and a tagging/release date column in POSIXct format.}

\item{id.groups}{Optional. A list containing ID groups, used to calculate stats independently
within each group, as well as comparing relationships between ids of different groups.}

\item{subset}{If defined, overlaps are calculated independently for each level
of this variable (usually a temporal class, e.g., diel phase or annual season).
If left NULL, single overlap indices are calculated for the whole monitoring period.}

\item{id.col}{Name of the column in id.metadata containing animal IDs. Defaults to 'ID'.}

\item{tagdate.col}{Name of the column in id.metadata containing animal tagging/release dates in
POSIXct format. Defaults to 'tagging_date'.}

\item{group.comparisons}{Controls the type of comparisons to be run, when id.groups are defined.
One of "within", "between" or "all". Useful to discard comparisons between individuals
belonging to the same group or skip comparisons between different groups, when these are
not required (less computing time). Defaults to "all".}
}
\value{
A list containing similarity matrices with pairwise overlaps in \%,
time windows in days (period used to compute overlap for each pair),
the average number of time-bins spent by each pair consecutively in the same receiver,
the maximum number of time-bins spent by each pair consecutively in the same receiver and
the total number of co-occurrences, as well as the number co-occurrences by location level.
If id.groups or a subset variable is defined, these results are nested within each of the grouping levels.
}
\description{
Estimates the extent of association between each two individuals,
assuming that joint space usage occurs whenever individuals overlap in
space (receiver) and time (bin). Only the total shared periods of detection
for a given pair are considered for this analysis, i.e. the time bins from the
latest release date to the earliest last detection between each pair. This
truncation step ensures that the absence of an individual in the receiver array
is not due to transmitter failure or premature death. This pairwise overlap
index is analogous to the simple ratio association index (Cairns & Schwager 1987,
Ginsberg & Young 1992), ranging from 0\% (no overlap) to 100\% (complete overlap).
}
