% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatePairDistances.R
\name{calculatePairDistances}
\alias{calculatePairDistances}
\title{Estimate minimum traveled distances between all positions (pairwise)}
\usage{
calculatePairDistances(
  data,
  land.shape,
  epsg.code = NULL,
  grid.resolution = 1000,
  mov.directions = 8,
  lon.col = "lon",
  lat.col = "lat",
  verbose = FALSE
)
}
\arguments{
\item{data}{A data frame with animal positions (COAs), containing longitude and latitude values}

\item{land.shape}{A  shape file containing coastlines.}

\item{epsg.code}{Coordinate reference system used to project positions (class 'CRS').
If not supplied, CRS is assumed to be the same as in land.shape.}

\item{grid.resolution}{Grid cell size in meters  over which shortest paths are going to be estimated.}

\item{mov.directions}{Number of directions allowed for shortest path calculation.
Passed to \link[gdistance]{transition}.}
}
\description{
Function to calculate the shortest path in water between consecutive positions.
If no land.shape is provided, linear ('great-circle') distances are calculated instead.
}
