% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{colorlegend}
\alias{colorlegend}
\title{Color Legend}
\usage{
colorlegend(
  col = femmecol(100),
  zlim,
  zlevels = 5,
  dz = NULL,
  zval = NULL,
  log = FALSE,
  posx = c(0.9, 0.93),
  posy = c(0.05, 0.9),
  main = NULL,
  main.cex = 1,
  main.col = "black",
  main.adj = 0.5,
  lab.col = "black",
  main.inset = 1,
  digit = 0,
  left = FALSE,
  lab.scientific = FALSE,
  ...
)
}
\description{
Creates a color legend for a plot, adapted from https://rdrr.io/cran/shape/src/R/colorlegend.R with additional features such as main.adj, main.inset, and support for scientific notation.
}
\note{
This function is intended for internal use within the 'moby' package.
}
\keyword{internal}
