% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{decimalPlaces}
\alias{decimalPlaces}
\title{Decimal Places}
\usage{
decimalPlaces(x)
}
\description{
Determines the number of decimal places in a numeric value.
}
\note{
This function is intended for internal use within the 'moby' package.
}
\keyword{internal}
