% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{filled.contour}
\alias{filled.contour}
\title{Filled Contour Plot with Modified Options}
\usage{
filled.contour(
  x = seq(0, 1, length.out = nrow(z)),
  y = seq(0, 1, length.out = ncol(z)),
  z,
  xlim = range(x, finite = TRUE),
  ylim = range(y, finite = TRUE),
  zlim = range(z, finite = TRUE),
  levels = pretty(zlim, nlevels),
  nlevels = 20,
  color.palette = cm.colors,
  col = color.palette(length(levels) - 1),
  plot.title,
  plot.axes,
  key.title,
  key.axes,
  asp = NA,
  xaxs = "i",
  yaxs = "i",
  las = 1,
  axes = TRUE,
  frame.plot = axes,
  mar,
  invert.scale = FALSE,
  ...
)
}
\description{
Produces a filled contour plot with additional modifications for flexibility.
}
\note{
This function is adapted from the original \code{filled.contour} function, modified by Ian Taylor, Carey McGilliard, and Bridget Ferris.
}
\keyword{internal}
