% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDefaults.R
\name{getDefaults}
\alias{getDefaults}
\title{Get Default Configuration Parameters}
\usage{
getDefaults(var)
}
\arguments{
\item{var}{A character string specifying the variable name for which to retrieve the default
setting. Accepted values are "id.col", "datetime.col", "timebin.col", "tagging.dates", and "tag.durations".
The first three can also be accessed using their abbreviated forms: "id", "datetime", and "timebin".}
}
\value{
The current default setting for the specified variable.
}
\description{
This function retrieves the current default parameters for the 'moby' package.
It accesses a dedicated environment where user-defined settings are stored, such as column
names for animal IDs, date-times, and time-bins.
}
\examples{
# Retrieve the default column name for animal IDs
getDefaults("id")

# Retrieve the default column name for date-times
getDefaults("datetime")

}
