% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDielPhase.R
\name{getDielPhase}
\alias{getDielPhase}
\title{Estimate diel phase}
\usage{
getDielPhase(datetimes, coords, phases = 2, solar.depth = 18)
}
\arguments{
\item{datetimes}{A POSIXct object containing the respective datetimes or time-bins.}

\item{coords}{A SpatialPoints, matrix, or data frame object containing geographic (unprojected)
longitude and latitude coordinates (in that order) for which to estimate sunrise and sunset times.
If a single point or a matrix/data frame with one row is provided, the same coordinates
will be used for all calculations.}

\item{phases}{Integer indicating the number of diel phases to return (2, 3, or 4).}

\item{solar.depth}{Numeric value indicating the angle of the sun below the horizon (in degrees).
Passed to the \code{\link[suntools]{crepuscule}} function.}
}
\value{
A factor indicating the diel phase.
}
\description{
Function to retrieve diel phase (e.g. day/night) for given datetimes and coordinates.\cr

The number of retrieved levels can be set by the 'phases' argument:\cr
\itemize{
\item phases=2:  day | night\cr
\item phases=3:  day | crepuscule | night\cr
\item phases=4:  dawn | day | dusk | night\cr
}

Crepuscular periods can be defined based on different solar elevation angles:\cr
\itemize{
\item solar.depth=6: civil twilight\cr
\item solar.depth=12: nautical twilight\cr
\item solar.depth=18: astronomical twilight\cr
}
}
\examples{
datetimes <- as.POSIXct("2024-05-30 12:00:00", tz = "UTC")
coords <- matrix(c(-7.997, 37.008), ncol = 2)
getDielPhase(datetimes, coords, phases = 4, solar.depth = 12)
}
