% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getPosition}
\alias{getPosition}
\title{Get Position}
\usage{
getPosition(keyword, inset)
}
\description{
Calculates the position of a keyword on a plot with optional inset adjustments.
}
\note{
This function is intended for internal use within the 'moby' package.
}
\keyword{internal}
