% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSeason.R
\name{getSeason}
\alias{getSeason}
\title{Estimate annual season}
\usage{
getSeason(datetimes, hemisphere = "Northern")
}
\arguments{
\item{datetimes}{A POSIXct object containing the respective datetimes or time-bins.}

\item{hemisphere}{A character string specifying the hemisphere ("Northern" or "Southern").}
}
\value{
A factor indicating the season.
}
\description{
This function determines the meteorological season for a given date based on the
meteorological definition, which divides the year into four seasons of three months each:\cr

Northern Hemisphere:\cr
\itemize{
\item Spring: March - May\cr
\item Summer: June - August\cr
\item Autumn: September - November\cr
\item Winter: December - February\cr
}

Southern Hemisphere:\cr
\itemize{
\item Spring: September - November\cr
\item Summer: December - February\cr
\item Autumn: March - May\cr
\item Winter: June - August\cr
}
}
\examples{
datetimes <- as.POSIXct("2024-05-30")
getSeason(datetimes, hemisphere="Northern")
}
