% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSunTimes.R
\name{getSunTimes}
\alias{getSunTimes}
\title{Retrieve diel phase boundary times (hours)}
\usage{
getSunTimes(
  sunriset.coords,
  start.time,
  end.time,
  by = "\%m",
  solar.depth = 18
)
}
\arguments{
\item{sunriset.coords}{A SpatialPoints, matrix or numeric object containing longitude and
latitude coordinates (in that order) at which to estimate sunrise and sunset times.}

\item{start.time}{A POSIXct object containing the earliest date of the monitoring period.}

\item{end.time}{A POSIXct object containing the latest date of the monitoring period.}

\item{by}{Date-time format (as defined by \code{\link[base]{strptime}}) containing the
time frame used to average sunrise, sunset and twilight times. Defaults to month ("\%m").}

\item{solar.depth}{Angle of the sun below the horizon (in degrees). Passed the
solarDep argument in \code{\link[suntools]{crepuscule}} function.
Defaults to 18 (astronomical twilight).}
}
\value{
Data frame with diel phase' boundary times (in hours)
}
\description{
Returns a table containing average sunrise, sunset and twilight times
for a given period (study duration).
'
}
