% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobyData.R
\name{is_moby}
\alias{is_moby}
\title{Check whether an object is a mobyData dataset}
\usage{
is_moby(x)
}
\arguments{
\item{x}{An object.}
}
\value{
A logical value.
}
\description{
Check whether an object is a mobyData dataset
}
\examples{
data(rays)
is_moby(rays)                # TRUE
is_moby(as.data.frame(rays)) # FALSE (plain data frame)
}
\seealso{
\code{\link{as_moby}}
}
