% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrationsTable.R
\name{migrationsTable}
\alias{migrationsTable}
\title{Create migrations table}
\usage{
migrationsTable(
  data,
  sites.col,
  aggregate.by = NULL,
  tag.dates = NULL,
  animals.info = NULL,
  id.groups = NULL,
  id.col = "ID",
  plot = T,
  plot.stats = "all",
  same.scale = T,
  von.bertalanffy = F,
  VBGF.params = NULL
)
}
\arguments{
\item{data}{A data frame containing binned animal detections.}

\item{sites.col}{Variable containing locations.}

\item{aggregate.by}{Variable containing locations.}

\item{tag.dates}{A data frame containing two columns, one containing the animal IDs
(column name defined by 'id.col') and the other the tagging dates in POSIXct format.}

\item{animals.info}{A data frame containing additional information about the tagged animals,
such as length, sex or transmitter type. If more than one row exists per animal,
variables are collapsed before being merged with the remaining stats.}

\item{id.groups}{Optional. A list containing ID groups, used to
visually aggregate animals belonging to the same class (e.g. different species or life stages).}

\item{id.col}{Name of the column containing animal IDs Defaults to 'ID'.}

\item{plot}{Boolean. If true a plot is generated containing all the stats together with temporal metrics.
If false, a summary data frame is returned.}

\item{plot.stats}{Indicates which variables to plot. Available options: "all", "none", "depart.hour",
"arrival.hour", "depart.month", "arrival.month", "duration". Defaults to "all".}

\item{same.scale}{Standardizes scales across plots and id groups.}

\item{von.bertalanffy}{Optional. Predict lengths at the moment of departure using a Von Bertalanffy Growth model.}

\item{VBGF.params}{A list containing the required parameters to fit the  Von Bertalanffy Growth model.
e.g.: 'Linf', 'K' and 't0'. Check \code{\link[TropFishR]{VBGF}} function for further details.}
}
\description{
Creates a table containing  migration metrics
(nº movements between sites, nº migrating individuals, etc.).
}
