% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobyNetwork.R
\name{mobyNetwork}
\alias{mobyNetwork}
\alias{is_mobyNetwork}
\alias{networkEdges}
\alias{networkNodes}
\alias{networkType}
\title{Inspect a moby network object}
\usage{
is_mobyNetwork(x)

networkEdges(x)

networkNodes(x)

networkType(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{is_mobyNetwork()} returns a logical; \code{networkEdges()} returns the edge data frame;
\code{networkNodes()} returns the node data frame; \code{networkType()} returns \code{"association"} or
\code{"movement"}.
}
\description{
Accessors and predicates for \code{mobyNetwork} objects (returned by
\code{\link{calculateAssociations}} and, in future, the movement-network functions).
A \code{mobyNetwork} is a \code{data.frame} of edges that additionally carries the node table and
network metadata as attributes.
}
\examples{
data(rays)
trans <- calculateTransitions(rays, spatial.col = "station")
is_mobyNetwork(trans)
networkType(trans)
head(networkEdges(trans))
head(networkNodes(trans))
}
\seealso{
\code{\link{calculateAssociations}}
}
