% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomizeOverlaps.R
\name{pairwiseOverlapLite}
\alias{pairwiseOverlapLite}
\title{Calculate pairwise spatio-temporal overlaps (lite version)}
\usage{
pairwiseOverlapLite(
  p,
  table,
  pairwise_combinations,
  complete_ids,
  id.groups,
  group.comparisons,
  start_dates,
  end_dates,
  metric
)
}
\description{
This internal function calculates the pairwise spatio-temporal overlaps
(association index) between two individuals. It is a streamlined adaptation of the internal
pairwiseOverlap function used within the calculateOverlap function. The function assumes
that joint space usage occurs whenever individuals overlap in space (receiver) and time (time bin).

This internal function calculates the pairwise spatio-temporal overlaps
(association index) between two individuals. It is a streamlined adaptation of the internal
\code{pairwiseOverlap} function used within the \code{\link{calculateOverlap}} function.
}
\note{
This function is intended for internal use within the 'moby' package.
}
\keyword{internal}
