% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBoxplots.R
\name{plotBoxplots}
\alias{plotBoxplots}
\title{Generates boxplots of multiple responses per grouping variable.}
\usage{
plotBoxplots(
  data,
  subsetted.overlap = NULL,
  subsetted.kuds = NULL,
  responses,
  by,
  id.col = getDefaults("id"),
  discard.incomplete = TRUE,
  display.n = FALSE,
  box.colors = NULL,
  outliers = FALSE,
  background.color = "gray96",
  box.pattern = FALSE,
  cols = 2
)
}
\arguments{
\item{data}{A data frame containing animal detections and distances traveled,
as returned by \code{\link{calculateTracks}}.}

\item{subsetted.overlap}{List containing pairwise overlaps estimated by a given
time-frame (e.g. diel phase), as returned by \code{\link{calculateOverlap}} using "subset".
Only required if overlap is included as a response.}

\item{subsetted.kuds}{List containing kernel utilization areas estimated by a given
time-frame (e.g. diel phase), as returned by \code{\link{calculateKUDs}} using "subset".
Only required if kud50 or kud95 are included as responses.}

\item{responses}{Metrics to include in the boxplots panel. Any of "detections",
"overlap", "total_dist, "kud50", "kud95" or any other included in the supplied data.}

\item{by}{Time-frame corresponding to the grouping category used to subset the data.}

\item{id.col}{Name of the column containing animal IDs. Defaults to 'ID'.}

\item{discard.incomplete}{Boolean to indicate if individuals with missing values (for any of the
grouping categories) should be discarded. If "by" contains only two grouping levels
(e.g. "day" and "night"), paired tests are computed whenever the individuals with data match between groups.}

\item{display.n}{If true, the number of individuals used to calculate each boxplot is
displayed together with the grouping labels.}

\item{box.colors}{Color of the box plot bodies.}

\item{outliers}{Logical. Specifies whether outliers should be included. Defaults to FALSE.}

\item{background.color}{Color for the plot background. Defaults to "gray96".}

\item{box.pattern}{Boolean to indicate if boxplot bodies should be filled with
different patterns.}

\item{cols}{Number of columns in the final panel (passed to the mfrow argument).}
}
\description{
Produces a panel with several boxplots (one per response), for
comparing multiple behavioural metrics across different grouping levels (usually time frames).
}
