% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDetections.R
\name{plotDetections}
\alias{plotDetections}
\title{Detection plots}
\usage{
plotDetections(
  binned.data,
  tagging.dates,
  id.groups = NULL,
  color.by = NULL,
  color.pal = NULL,
  date.format = "\%b/\%y",
  date.interval = 4,
  date.start = 1,
  sunriset.coords,
  id.col = "ID",
  discard.missing = T,
  background.color = "gray96",
  highlight.isolated = T,
  pt.cex = 1.5,
  crepuscular.lines = F,
  grid = F,
  grid.color = "white",
  cols = 2,
  legend.cols = 3
)
}
\arguments{
\item{binned.data}{A data frame containing animal detections with corresponding time-bins (must contain a 'timebin' column).}

\item{tagging.dates}{A POSIXct vector containing the tag/release date of each animal.}

\item{id.groups}{Optional. A list containing ID groups, used to
visually aggregate animals belonging to the same class (e.g. different species).}

\item{color.by}{Optional. Variable defining the color group of individual detections.
Can be used for example to depict detections of different receivers or in different habitats.}

\item{color.pal}{Color palette for the points.}

\item{date.format}{Date-time format (as used in \code{\link[base]{strptime}}),
defining the x-axis labels. Defaults to month-year ("\%b/\%y").}

\item{date.interval}{Number defining the interval between each
displayed date (x-axis label). Defaults to 4.}

\item{date.start}{Integer defining the first displayed date (can be used in combination
with 'date.interval" to better control the x-axis labels). Defaults to 1.}

\item{sunriset.coords}{A SpatialPoints or matrix object containing longitude and
latitude coordinates (in that order) at which to estimate sunrise and sunset times.}

\item{id.col}{Name of the column containing animal IDs. Defaults to 'ID'.}

\item{discard.missing}{If true, only individuals with detections are included.
If false, empty plots are drawn for missing individuals.}

\item{background.color}{Color for the plot background. Defaults to "gray96".}

\item{highlight.isolated}{If a 'color.by' variable is defined, detections are ordered
and plotted according to their "density", with isolated detections being brought forward
to prevent them from being hidden behind points' clouds. Defaults to True.}

\item{pt.cex}{Expansion factor(s) for the points (detections). Defaults to 1.5.}

\item{crepuscular.lines}{Boolean to indicate if "dawn" and "dusk" lines should be plotted. Defaults to False.}

\item{grid}{If true, a grid is plotted (horizontal and vertical guides across the entire plot). Defaults to False.}

\item{grid.color}{Color of the grid lines. Defaults to white.}

\item{cols}{Number of columns in the final panel (passed to the mfrow argument). Defaults to 2.}

\item{legend.cols}{Number of columns in the legend (when 'color.by' is defined).  Defaults to 3.}
}
\description{
Plots color-coded detections over time and date, independently for each individual.
}
