% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGroupSizes.R
\name{plotGroupSizes}
\alias{plotGroupSizes}
\title{Plot frequency distributions for co-occurring group sizes}
\usage{
plotGroupSizes(
  table,
  id.groups = NULL,
  split.by = NULL,
  group.comparisons = "all",
  bar.colors = NULL,
  levels.order = NULL,
  title.cex = 1.1,
  label.cex = 1,
  axis.cex = 0.9,
  annot.cex = 0.7,
  legend.pos = "topright",
  legend.style = "vertical",
  legend.cex = 0.7,
  cols = 1
)
}
\arguments{
\item{table}{A data frame containing binned detections in the wide format
(time bin x individual matrix, with values corresponding to the receiver/station with the
highest number of detections), as returned by \code{\link{createWideTable}}.}

\item{id.groups}{Optional. A list containing ID groups, used to generate different
plots to each class (e.g. different species).}

\item{split.by}{Optional. If defined, frequencies are calculated separately for each level
of this variable. If NULL, frequencies are calculated for the entire study duration.}

\item{group.comparisons}{Controls the type of comparisons to be run, when id.groups are defined.
One of "within", "between" or "all". Defaults to "all".}

\item{bar.colors}{A vector of colors for the bars. Nº of colors should match
the nº of levels in the 'split.by' variable.}

\item{levels.order}{Numeric vector with the preferred order of the levels defined by the 'split.by' variable
(in the stacked barplot).}

\item{title.cex}{Numeric. Font size for the plot title. Defaults to 1.1.}

\item{label.cex}{Numeric. Font size for axis labels. Defaults to 1.}

\item{axis.cex}{Numeric. Font size for axis text. Defaults to 0.9.}

\item{annot.cex}{Numeric. Font size for bar annotation text. Defaults to 0.7.}

\item{legend.pos}{Legend position.}

\item{legend.style}{Orientation of the legend, either "horizontal" or "vertical".}

\item{legend.cex}{Numeric. Font size for legend text. Defaults to 0.7.}

\item{cols}{Number of columns in the final panel (passed to the mfrow argument).}
}
\description{
Function to calculate the frequency distribution of the
nº of animals co-occurring at each time-bin, either for the
entire study duration or split by a given variable (e.g. diel phase).
}
