% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOverlap.R
\name{plotOverlap}
\alias{plotOverlap}
\title{Network representation of pairwise overlaps.}
\usage{
plotOverlap(
  overlaps = NULL,
  random.results = NULL,
  color.nodes.by = "group",
  scale.nodes.by = NULL,
  remove.missing = FALSE,
  min.val = NULL,
  cut.val = NULL,
  group.order = NULL,
  plot.stats = TRUE,
  cex.title = 1.4,
  cex.lab = 1.2,
  cex.axis = 1,
  cex.legend = 1,
  legend.inset = c(-0.02, 0),
  network.layout = "spring",
  nodes.color = NULL,
  nodes.size = c(1.2, 2.2),
  nodes.label.cex = 1,
  nodes.label.scale = FALSE,
  nodes.label.color = "white",
  edge.color = NULL,
  edge.curved = 0.5,
  edge.width = c(0.4, 3.5),
  edge.label.cex = 1.4,
  edge.label.color = "black",
  edge.label.font = 1,
  background.col = "grey96",
  overlap.line.color = "red2",
  overlap.line.lwd = 3,
  overlap.line.lty = 1,
  hist.side = "bottom",
  standardize.freqs = FALSE,
  cols = NULL,
  ...
)
}
\arguments{
\item{overlaps}{A data frame containing the pairwise overlap data. Should be the output from the \code{calculateOverlap} function.}

\item{random.results}{A list containing null model permutation results. Should be the output from the \code{randomizeOverlaps} function.}

\item{color.nodes.by}{Variable used to color nodes (individuals). It should be a categorical vector (factor or character).
Alternatively, it can be set to 'group' (default) to assign different colors to different ID groups.}

\item{scale.nodes.by}{Numeric vector to scale the nodes. Length should match the number of IDs in overlaps.}

\item{remove.missing}{Logical. If TRUE, excludes individuals without detections and null pairwise comparisons from the network.
Defaults to FALSE.}

\item{min.val}{Minimum value for overlap threshold. Defaults to the 50\% quantile of the overlap values.}

\item{cut.val}{Cutoff value for overlap threshold. Defaults to the 90\% quantile of the overlap values.}

\item{group.order}{Order of groups for plotting. Defaults to NULL.}

\item{plot.stats}{Logical. If TRUE, additional network metrics are plotted
below each network. Defaults to TRUE.}

\item{cex.title}{Size of the plot title. Defaults to 1.4.}

\item{cex.lab}{Size of the axis labels. Defaults to 1.2.}

\item{cex.axis}{Size of the axis text. Defaults to 1.0.}

\item{cex.legend}{Size of the legend text. Defaults to 1.0.}

\item{legend.inset}{Optional. Specifies how far the legend is inset from the
null model histogram margins. If a single value is given, it is used for both margins;
if two values are given, the first is used for x- distance, the second for y-distance.
Defaults to c(-0.02, 0).}

\item{network.layout}{Layout algorithm for the network (see \code{\link[qgraph]{qgraph}}).
Defaults to "spring".}

\item{nodes.color}{Color(s) for the nodes.}

\item{nodes.size}{Numeric vector of length 2, indicating the minimum and maximum
node sizes. Defaults to c(1, 2.2).}

\item{nodes.label.cex}{Font size for node labels. Defaults to 0.8.}

\item{nodes.label.scale}{Logical. If TRUE, scales node labels. Defaults to FALSE.}

\item{nodes.label.color}{Color of the node labels. Defaults to white.}

\item{edge.color}{Color(s) for the edges. Defaults to NULL, which averages node colors.}

\item{edge.curved}{Numeric value or logical indicating the curvature of the edges. Defaults to 0.5.}

\item{edge.width}{Numeric vector of length 2, indicating the minimum and maximum edge widths. Defaults to c(0.4, 3.5).}

\item{edge.label.cex}{Font size for edge labels. Defaults to 1.}

\item{edge.label.color}{Color of the edge labels. Defaults to black.}

\item{edge.label.font}{Font type for edge labels (1 is plain, 2 is bold, 3 is italic, 4 is bold and italic, 5 is symbol font). Defaults to 1.}

\item{background.col}{Background color of the plot. Defaults to "grey96".}

\item{overlap.line.color}{Color of the line representing observed overlap in the null model plot. Defaults to "red2".}

\item{overlap.line.lwd}{Line width of the observed overlap line. Defaults to 3.}

\item{overlap.line.lty}{Line type of the observed overlap line. Defaults to 1.}

\item{hist.side}{Position of the null model histogram relative to the network plot, either "bottom" or "left". Defaults to "bottom".}

\item{standardize.freqs}{Logical. If TRUE, standardizes frequencies in the null model plot. Defaults to FALSE.}

\item{cols}{Number of columns for the plot layout. Defaults to NULL.}

\item{...}{Further arguments passed to \code{\link[qgraph]{qgraph}}.}
}
\description{
This function plots a network of estimated pairwise overlaps between
individuals or a histogram showing the distribution of overlaps from null model permutations,
or both. If only \code{overlaps} are provided, it plots a network where each node represents
an individual, with edges indicating the extent of spatial overlap between pairs.
Thicker edges represent higher overlap. If only \code{random.results} are provided, it plots
a histogram showing the distribution of overlaps from the null model permutations.
This helps to identify whether the observed overlaps are significantly different from
what would be expected by chance. If both are provided, it plots both the network
and the histogram.
}
\seealso{
\code{calculateOverlap}, \code{randomizeOverlaps}
}
