% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOverlapFrequences.R
\name{plotOverlapFrequences}
\alias{plotOverlapFrequences}
\title{Plot frequencies of co-occurring group sizes}
\usage{
plotOverlapFrequences(
  table,
  n.individuals,
  split.by = NULL,
  id.groups = NULL,
  group.comparisons = "all",
  legend.pos = "topright",
  legend.style = "horizontal",
  bar.colors = NULL,
  levels.order = NULL,
  cols = 1
)
}
\arguments{
\item{table}{A data frame object containing binned detections in the wide format
(time bin x individual matrix, with values corresponding to the receiver with the
highest number of detections), as returned by \code{\link{createWideTable}}.}

\item{n.individuals}{Number of tagged individuals.}

\item{split.by}{Optional. If defined, frequencies are calculated separately for each level
of this variable. If NULL, frequencies are calculated for the entire study duration.}

\item{id.groups}{Optional. A list containing ID groups, used to generate different
plots to each class (e.g. different species).}

\item{group.comparisons}{Controls the type of comparisons to be run, when id.groups are defined.
One of "within", "between" or "all". Defaults to "all".}

\item{legend.pos}{Legend position.}

\item{legend.style}{Legend style (horizontal vs vertical).}

\item{bar.colors}{A vector of colors for the bars. Nº of colors should match
the nº of levels in the 'split.by' variable.}

\item{levels.order}{Numeric vector with the preferred order of the levels defined by the 'split.by' variable
(in the stacked barplot).}

\item{cols}{Number of columns in the final panel (passed to the mfrow argument).}
}
\description{
Function to calculate the frequency distribution of the
nº of animals co-occurring at each time-bin, either for the
entire study duration or split by a given variable (e.g. diel phase).
}
\seealso{
\code{\link{calculateOverlap}}
}
