% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOverlapNetwork.R
\name{plotOverlapNetwork}
\alias{plotOverlapNetwork}
\title{Network representation of pairwise overlaps.}
\usage{
plotOverlapNetwork(
  overlaps,
  random.results = NULL,
  id.metadata,
  id.col = getDefaults("id"),
  color.by = NULL,
  scale.by = NULL,
  nodes.size = c(0.7, 1.4),
  label.cex = 1.2,
  nodes.color = ("#643BA6"),
  edge.label.cex = 1.7,
  min.val = NULL,
  cut.val = NULL,
  color.pal = NULL,
  ...
)
}
\arguments{
\item{overlaps}{Similarity matrix containing pairwise overlaps, as returned by \code{\link{calculateOverlap}}.}

\item{id.metadata}{A data frame containing animal IDs as well as an optional 'color.by' and 'size.by' variables.}

\item{id.col}{Name of the column containing animal IDs. Defaults to 'ID'.}

\item{color.by}{Optional. Factor variable in id.metadata defining the color of each node (e.g. "species", "sex", etc).}

\item{scale.by}{Optional. Numeric variable in id.metadata defining the size of each node (e.g.  "length").}

\item{min.val}{Edges with overlaps lower than this value are not shown.}

\item{cut.val}{Edges with overlaps larger than this value will have the strongest color intensity
and become wider the stronger they are, and edges with overlaps lower than this value
will have the smallest width and become vaguer the weaker the weight.}

\item{color.pal}{Color of the nodes.}

\item{...}{Further arguments passed to \code{\link[qgraph]{qgraph}}.}

\item{model}{Type of regression in case animal.sizes have been supplied ("linear" or "quadratic").
A vector with two elements should be supplied, the first to be used in the "average overlap ~ size" graph,
and the second in the "pairwise overlap ~ size difference" graph.}
}
\description{
Function to plot a network of estimated pairwise overlaps.
Each node represents a single individual and can be dimensioned proportionally
to its size. Edges represent the extent of spatial overlap between each pair
(i.e. greater thickness indicates a higher overlap). If animals sizes are supplied,
two additional plots are produced, one representing the relationship between
animals' size/length and average overlap and the other representing the
relationship between pairwise differences in size and overlap scores.
}
\seealso{
\code{\link{calculateOverlap}}

\code{\link{plotOverlapFrequences}}
}
