% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOverlapTable.R
\name{plotOverlapTable}
\alias{plotOverlapTable}
\title{Generate a visual representation of pairwise overlap statistics.}
\usage{
plotOverlapTable(
  random.results,
  type = "significance",
  color.pal = NULL,
  sort.by = NULL,
  sort.by.title = NULL,
  full.scale = FALSE,
  discard.missing = TRUE,
  table.cex = 0.8,
  legend.cex = 0.7
)
}
\arguments{
\item{random.results}{A list containing null model permutation results. Should be the output from the \code{\link{randomizeOverlap}} function.}

\item{type}{The type of metric to be plotted. Must be one of "significance" (default) or "mean overlap".}

\item{color.pal}{A vector of colors defining the palette for the table.
\itemize{
\item If \code{type = "significance"}, the palette should include three colors
corresponding to positive ("+"), negative ("-"), and non-significant ("ns") results.
\item For \code{type = "mean overlap"}, a gradient of at least 10 colors is recommended.
}}

\item{sort.by}{An optional vector to reorder IDs (e.g., based on their size or
another quantitative metric). Must have the same length as the IDs in \code{random.results}.}

\item{sort.by.title}{A string providing a title for the \code{sort.by} metric
(e.g., "Size (cm)"). If not supplied, the function attempts to extract the name
of the variable passed to \code{sort.by}.}

\item{full.scale}{Logical. If \code{TRUE} and \code{type = "mean overlap"},
scales overlap values from 0\% to 100\%. Defaults to \code{FALSE}.}

\item{discard.missing}{Logical. If \code{TRUE}, removes rows and columns with
no valid pairwise interactions. Defaults to \code{TRUE}.}

\item{table.cex}{Numeric. Character expansion factor for the table text.
Defaults to \code{0.7}.}

\item{legend.cex}{Numeric. Character expansion factor for the legend text.
Defaults to \code{0.7}.}
}
\description{
Plots a contingency table illustrating the pairwise distribution of overlap
scores/significance across individuals. IDs can be sorted by length or any other quantitative metric.
}
\seealso{
\code{\link{calculateOverlap}} for overlap computation and
\code{\link{randomizeOverlap}} for generating \code{random.results}.
}
