% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRandomOverlaps.R
\name{plotRandomOverlaps}
\alias{plotRandomOverlaps}
\title{Plot overlap randomization results}
\usage{
plotRandomOverlaps(
  overlaps,
  random.results,
  include.table = F,
  type = "mean",
  conf.level = 0.95
)
}
\arguments{
\item{overlaps}{Similarity matrix containing pairwise overlaps, as returned by \code{\link{calculateOverlap}}.}

\item{random.results}{Overlap simulations, as returned by \code{\link{randomizeOverlaps}}.}

\item{include.table}{Boolean to indicate if a summary table should be drawn below the plot.}

\item{type}{Type of statistic used to plot the histogram and calculate overlap significance.
One of "mean" or "median". Defaults to "mean".}

\item{conf.level}{Confidence level. Defaults to 95\%.}
}
\description{
Plots frequency distributions of mean/median overlaps
obtained through the Monte Carlo permutation tests. Null hypotheses are tested
by comparing the observed mean/median overlaps (Obs. overlap, indicated by an arrow)
to the upper and lower tails of the null distributions, using a chosen confidence level.
}
