% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRetention.R
\name{plotRetention}
\alias{plotRetention}
\title{Plot retention curve(s)}
\usage{
plotRetention(
  data,
  id.metadata,
  site.col = "habitat",
  id.col = "ID",
  tagdate.col = "tagging_date",
  tagsite.col = "tagging_location",
  color.by = NULL,
  color.pal = NULL,
  aggregate.fun = "mean",
  same.scale = F,
  background.col = "grey96",
  cols = 1,
  von.bertalanffy = F,
  VBGF.params = NULL
)
}
\arguments{
\item{data}{A data frame containing binned animal detections.}

\item{id.metadata}{A data frame containing the tagging location of each animal ('tagging_location' column),
as well as an optional (continuous) variable to generate a color scale.}

\item{site.col}{Should match the name of a variable in the supplied data containing
similar categories as those used in the tagging sites (e.g. pointing to specific receivers or broader areas/locations).
Defaults to "habitat".}

\item{id.col}{Name of the column containing animal identifications. Defaults to "ID".}

\item{tagdate.col}{Name of the column in id.metadata containing animal tagging/release dates in
POSIXct format. Defaults to 'tagging_date'.}

\item{tagsite.col}{Name of the column in id.metadata containing animal tagging/release locations.
Defaults to 'tagging_location'.}

\item{color.by}{Optional. Variable contained in the id.metadata table, used to generate
a color gradient (e.g. length).}

\item{color.pal}{Optional. Color palette used for the color gradient.}

\item{aggregate.fun}{Function used to aggregate numeric values if a color.by argument is supplied.
Defaults to "mean".}

\item{same.scale}{Logical. If TRUE, uses the same color scale for all panels. Defaults to FALSE.}

\item{background.col}{Background color of the plot. Defaults to "gray96".}

\item{cols}{Number of columns in the final panel (passed to the mfrow argument).}

\item{von.bertalanffy}{Optional. In case 'color.by' represents individual lengths measured at tagging,
it predicts new lengths at each time frame based on a Von Bertalanffy Growth model.}

\item{VBGF.params}{A list containing the required parameters to fit the  Von Bertalanffy Growth model.
e.g.: 'Linf', 'K' and 't0'. Check \code{\link[TropFishR]{VBGF}} function for further details.}
}
\description{
Function used to plot retention rate(s) at each one of the tagging sites.
It calculates the percentage of individuals that occur within their original tagging location over
a given time frame, relatively to the total number of released individuals. The algorithm considers the last detection
of each animal in that particular site (independently of the presences/absence periods or
migrations in-between).
}
