% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStationsMap.R
\name{plotStationsMap}
\alias{plotStationsMap}
\title{Plot map with receiver locations}
\usage{
plotStationsMap(
  data,
  split.by = NULL,
  id.col = getDefaults("ID"),
  lon.col = getDefaults("lon"),
  lat.col = getDefaults("lat"),
  pie.color = NULL,
  land.shape,
  land.color = "gray50",
  background.col = "#F3F7F7",
  background.layer = NULL,
  background.pal = NULL,
  scale.meters = NULL,
  scale.pos = "bottomright",
  scale.inset = 0.1
)
}
\arguments{
\item{data}{A data frame containing animal detections with corresponding time-bins.}

\item{split.by}{If defined, detection frequencies on each receiver are calculated
separately for each level of this variable. If NULL, frequencies are calculated
for the entire study duration. Defaults to "timeofday"..}

\item{id.col}{Name of the column containing animal IDs. Defaults to 'ID'.}

\item{lon.col}{Name of the column containing longitude values. Defaults to 'lon'.}

\item{lat.col}{Name of the column containing latitude values. Defaults to 'lat'.}

\item{pie.color}{Color of the pie charts. Single value if 'split.by' is not defined, otherwise
a vector containing the same number of colors as the number of levels in the grouping variable.}

\item{land.shape}{A shape file containing coastlines.}

\item{land.color}{Color of land areas.}

\item{background.col}{Background color if a background layer is not supplied.}

\item{background.layer}{A projected raster containing a variable to be displayed in the background
(e.g. bathymetry, temperature, etc.).}

\item{background.pal}{Color palette for the background layer.}

\item{scale.meters}{Distance covered by the scale bar, in meters. If null, it is
automatically defined as 20\% of the plot region.}

\item{scale.pos}{Position of the scale bar, specified by keyword.}

\item{scale.inset}{Inset distance(s) from the margins as a fraction of the
plot region, relatively to the scale bar position.}
}
\description{
Function to plot all receivers locations on a map. Pie charts representing the
average of detection frequencies per receiver are displayed above. Additionally,
detection frequencies can be calculated separately by a grouping variable (e.g. diel phase).
}
