% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTemporalKUDs.R
\name{plotTemporalKUDs}
\alias{plotTemporalKUDs}
\title{Plot KUD time series}
\usage{
plotTemporalKUDs(
  subsetted.kuds,
  label,
  id.groups = NULL,
  plot.by = "group",
  shade.season = F,
  same.scale = T
)
}
\arguments{
\item{subsetted.kuds}{List containing kernel utilization areas estimated by
time-frame (e.g. month), as returned by \code{\link{calculateKUDs} using "subset"}}

\item{label}{Time-frame used to subset KUDs, displayed on the x-axis (e.g. "Month").}

\item{id.groups}{Optional. A list containing ID groups, used to
visually aggregate animals belonging to the same class (e.g. different species).}

\item{plot.by}{Either "group" or "kud". If "group", the function plots a graph by id.group (overlaying KUD 50\% and KUD95\%).
If by "kud" the function plots a graph by KUD percentage (overlaying group series).}

\item{shade.season}{Boolean to indicate if seasons should be depicted in the background
(useful when plotting KUDs by month). Defaults to false.}
}
\description{
Function to plot the variation of 95\% and 50\% kernel utilization areas
across a given time-frame (usually month).
}
\seealso{
\code{\link{calculateKUDs}}
}
