% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{printConsole}
\alias{printConsole}
\title{Print to console}
\usage{
printConsole(string)
}
\arguments{
\item{string}{A character string to be printed to the console.}
}
\description{
Prints a string to the console with a specific formatting.
}
\note{
This function is intended for internal use within the \code{moby} package.
}
\keyword{internal}
