% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomizeOverlaps.R
\name{randomize}
\alias{randomize}
\title{Randomize detections}
\usage{
randomize(table, id.cols, start_dates, end_dates)
}
\description{
This function randomizes the detection data while preserving the structure within
constraint groups.
}
\note{
This function is intended for internal use within the 'moby' package.
}
\keyword{internal}
