% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomizeOverlap.R
\name{randomizeOverlap}
\alias{randomizeOverlap}
\title{Test significance of estimated overlaps through null model randomization tests}
\usage{
randomizeOverlap(
  table,
  overlaps,
  constraint.by = NULL,
  iterations = 1000,
  alternative = c("two.sided"),
  conf.level = 0.95,
  cores = 1,
  random.seed = NULL
)
}
\arguments{
\item{table}{A data frame containing binned detections in the wide format
(time bin x individual matrix, with values corresponding to the receiver/station with the
highest number of detections), as returned by \code{\link{createWideTable}}.}

\item{overlaps}{Data frame containing paiwise overlaps, as returned by \code{\link{calculateOverlap}}.}

\item{constraint.by}{Optional. Variable(s) to constrain permutations, e.g. to account
for potential diel or seasonal trends in animal occurrences. If supplied, permutations
across time bins will be restricted to the same combination of levels of these variables.}

\item{iterations}{Number of Monte Carlo iterations (simulated datasets). Defaults to 1000.}

\item{alternative}{Character string specifying the alternative hypothesis. Must be one of "two.sided",
"less", or "greater". "two.sided" tests for deviation in either direction, "less" tests if the observed
value is significantly less than the null distribution, and "greater" tests if the observed value is
significantly greater than the null distribution. Defaults to "two.sided".}

\item{conf.level}{Confidence level for the test. Defaults to 0.95 (95\%).}

\item{cores}{Number of CPU cores to use for the computations. Defaults to 1, which
means no parallel computing (single core).  If set to a value greater than 1,
the function will use parallel computing to speed up calculations.
Run \code{parallel::detectCores()} to check the number of available cores.}

\item{random.seed}{Optional. Set the seed for a reproducible randomization.
See \code{\link[base]{set.seed}}.}
}
\value{
A list containing:
\describe{
\item{\strong{summary}}{A summary table with overall results.}
\item{\strong{pairwise_results}}{A data frame similar to that returned by the \code{\link{calculateOverlap}}
function, but with additional columns indicating the mean of the null distribution,
empirical p-values, and association type for each animal dyad.}
\item{\strong{randomized_overlaps}}{A list containing all simulated pairwise overlaps
for each iteration. If a subset variable was specified in the
supplied overlaps, this will contain a nested list, where the first level will
contain an element for each level of the subset variable and the second level an
element containing all simulated pairwise overlaps for each iteration.}
}\cr

The summary table includes the following columns:
\itemize{
\item {Type}: The type of comparison. This column is included only when \code{id.groups} are identified in the supplied overlaps.
\item {Subset}: This column (e.g., a temporal subset) is included only when a subset variable was used to calculate the supplied overlaps.
\item {N dyads}: The number of dyads (pairs of individuals).
\item {Mean interval (d)}: The mean monitoring period (in days).
\item {Mean overlap (\%)}: The mean observed overlap percentage with standard deviation.
\item {Mean null distr (\%)}: The mean overlap percentage from the null distribution with standard deviation.
\item {P-value}: The estimated p-value(s), indicating the significance of the observed overlaps.
\item {Association}: The association type (e.g., positive, negative, non-significant) based on the p-value.
\item {Pairs Non Sig}: The number of pairs with non-significant.
\item {Pairs > Random}: The number of pairs with observed overlap greater than the null distribution.
\item {Pairs < Random}: The number of pairs with observed overlap less than the null distribution.
}
}
\description{
This function tests the null hypothesis of temporally independent space use
(i.e., each animal occurs independently of the other) using Monte Carlo permutation tests.
The algorithm permutes entries within each column, keeping the total number of detections
of each individual and the relative occurrence frequencies across receivers unchanged.
Permutations can be constrained by given factors (e.g., diel phase) and are performed
only within the monitoring period of each animal (interval between its release
and the date of its last detection). Empirical p-values are calculated by comparing
the observed spatiotemporal overlaps against their null distribution. The p-values can
be one-tailed or two-tailed based on the specified alternative hypothesis (a
continuity correction is applied to avoid p-values of exactly 0 or 1).
}
\references{
Gotelli, N. J. (2000). Null model analysis of species co-occurrence patterns. Ecology, 81(9), 2606-2621.\cr\cr
Farine, D. R. (2017). A guide to null models for animal social network analysis. Methods Ecol Evol 8: 1309–1320.
}
\seealso{
\code{\link{calculateOverlap}}, \code{\link{plotOverlap}}
}
