% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{rescale}
\alias{rescale}
\title{Rescale}
\usage{
rescale(x, to = c(0, 1), from = range(x, na.rm = TRUE, finite = TRUE), ...)
}
\description{
Rescales a numeric vector to a specified range.
}
\note{
This function is intended for internal use within the 'moby' package.
}
\keyword{internal}
