% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{rescale_vertex_igraph}
\alias{rescale_vertex_igraph}
\title{Rescale Vertex Igraph}
\usage{
rescale_vertex_igraph(
  vertex.size,
  par.usr = par("usr"),
  minmax.relative.size = c(0.01, 0.04),
  adjust = 200
)
}
\arguments{
\item{vertex.size}{Numeric vector of unscaled vertices' sizes. This is unit-free.}

\item{par.usr}{Integer vector of length 4 with the coordinates of plotting region.
by default uses \code{par("usr")}.}

\item{minmax.relative.size}{A numeric vector of length 2. Represents the
desired min and max vertex sizes relative to the x-axis in terms of percentage
(see details).}

\item{adjust}{Numeric scalar. Adjustment made to the resulting adjusted size
(see details).}
}
\description{
Rescale vertex size to be used in \code{\link[igraph:plot.igraph]{plot.igraph}}.
This function rescales a vertex size before passing it to \code{\link[igraph:plot.igraph]{plot.igraph}}
so that the resulting vertices have the desired size relative to the x-axis.
}
\note{
This function is intended for internal use within the 'moby' package.
}
\author{
George G. Vega Yon
}
\keyword{internal}
