% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{scalebar}
\alias{scalebar}
\title{Scale Bar}
\usage{
scalebar(
  d,
  xy = NULL,
  type = "line",
  divs = 2,
  below = "",
  bar.lwd = 0.4,
  bar.height = 1.5,
  lonlat = NULL,
  label,
  adj = c(0.5, -0.5),
  lwd = 2,
  ...
)
}
\description{
Adds a scale bar to a plot, supporting both line and bar types.
}
\note{
This function is intended for internal use within the 'moby' package.
}
\keyword{internal}
