% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryTable v2.R
\name{summaryTable2}
\alias{summaryTable2}
\title{Generate summary table for tagged animals}
\usage{
summaryTable2(
  data,
  tagging.dates = getDefaults("tagdates"),
  tag.durations = NULL,
  id.metadata = NULL,
  id.col = getDefaults("id"),
  datetime.col = getDefaults("datetime"),
  station.col = getDefaults("station"),
  residency.index = "IR1",
  residency.by = NULL,
  id.groups = NULL,
  error.stat = "sd"
)
}
\arguments{
\item{data}{A data frame containing animal detections. Each row should represent an individual detection event,
unless a 'detections' column is included to indicate the number of detections for each row.}

\item{tagging.dates}{A POSIXct vector containing the tag/release date of each animal.
The length of this vector should match the number of unique animal IDs.
Alternatively, if a single value is provided, it will be applied to all IDs.}

\item{tag.durations}{Optional. A numeric vector containing the estimated battery
duration of the deployed tags (in days). If a single value is provided, it will be applied to all IDs.}

\item{id.metadata}{A data frame containing metadata about the tagged animals, such as their length,
sex, or transmitter type. If there are multiple rows per animal, variables will be collapsed before merging with other statistics.}

\item{id.col}{Name of the column containing animal IDs. Defaults to 'ID'.}

\item{datetime.col}{Name of the column containing datetimes in POSIXct format. Defaults to 'datetime'.}

\item{station.col}{Name of the column containing station/receiver IDs. Defaults to 'station'.}

\item{residency.index}{A character string specifying the type of residency index to calculate.
Options include:
\itemize{
\item "IR1": Residency Index 1, calculated as the number of days the animal was detected (Dd) divided by the detection interval (Di), i.e., the number of days between tagging and last detection (days at liberty). This represents a maximum residency value, considering only the period for which the animal was known to be alive and the tag operational.
\item "IR2": Residency Index 2, calculated as the number of days the animal was detected (Dd) divided by the study interval (Dt), i.e., the total number of days between release and last data download or tag expiration date. This approach provides a minimum residency value, assuming the animal was alive and detectable throughout the study period.
\item "IWR": Weighted Residency Index, which combines two ratios: Dd/Dt (days detected over study interval) weighted by Di/Dt (detection interval over study interval). It provides a more nuanced measure of residency that balances the time an animal is detected with the total study period.
}

The choice of index can affect the interpretation of residency patterns, so it's important to select the one that best fits the study objectives.
Defaults to "IR1". If \code{tag.durations} are provided, the default changes to "IR2".
Further information on residency estimation methods can be found in Kraft et al. (2023) - see below in the references section.}

\item{residency.by}{Optional. Variable used to calculate partial residencies (e.g. array or habitat).
Defaults to NULL.}

\item{id.groups}{Optional. A list where each element is a group of IDs, used for visually aggregating
animals belonging to the same class (e.g., different species or life stages). If supplied, averages
will be calculated independently for each group.}

\item{error.stat}{The statistic to use for variability/error calculation, either 'sd' (standard deviation)
or 'se' (standard error). Defaults to 'sd'.}
}
\description{
This function generates a summary table with information about tagged animals,
including tagging dates, last detection dates, and various monitoring and residency metrics. It allows for the
inclusion of additional metadata and the calculation of overall means and error metrics.
}
\references{
Kraft, S., Gandra, M., Lennox, R. J., Mourier, J., Winkler, A. C., & Abecasis, D. (2023).
Residency and space use estimation methods based on passive acoustic telemetry data.
Movement Ecology, 11(1), 12.
https://doi.org/10.1186/s40462-023-00349-y
}
