% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateCOAs.R
\name{calculateCOAs}
\alias{calculateCOAs}
\title{Calculate Centers of Activity (COAs)}
\usage{
calculateCOAs(
  data,
  id.col = getDefaults("ID"),
  timebin.col = getDefaults("timebin"),
  lon.col = getDefaults("lon"),
  lat.col = getDefaults("lat"),
  station.col = NULL
)
}
\arguments{
\item{data}{A data frame containing animal detections and including a time bin column
(as specified by the \code{timebin.col} argument). Time bins can be created using the
\code{\link{getTimeBins}} function.}

\item{id.col}{Name of the column containing animal IDs. Defaults to 'ID'.}

\item{timebin.col}{Name of the column containing time bins (in POSIXct format). Defaults to 'timebin'.}

\item{lon.col}{Name of the column containing longitude values. Defaults to 'lon'. Coordinates can be supplied
either in geographic (unprojected) or projected format. In both cases, the appropriate projected CRS
should be provided via the \code{epsg.code} argument.}

\item{lat.col}{Name of the column containing latitude values. Defaults to 'lat'. Coordinates can be supplied
either in geographic (unprojected) or projected format. In both cases, the appropriate projected CRS
should be provided via the \code{epsg.code} argument.}

\item{station.col}{Name of the column containing station/receiver IDs. Defaults to 'station'.}
}
\value{
A data frame containing the following columns:
\itemize{
\item The unique identifier (\code{id.col}) for each individual.
\item The time bin (\code{timebin.col}) for which the COA was calculated.
\item Mean longitude and latitude (\code{lon.col} and \code{lat.col}) for each ID and time bin.
\item The number of detections (\code{detections}) for each ID and time bin.
\item The number of unique stations visited (\code{stations}) for each ID and time bin.
\item For numeric columns: Mean values for each ID and time bin.
\item For character or factor columns: Concatenated unique values (separated by "|") for each ID and time bin.
}
}
\description{
This function calculates position estimates (Centers of Activity, COAs)
from presence data collected at multiple receivers. It uses weighted means of longitude
and latitude based on detection counts during specified time bins. Additionally, it aggregates
all remaining columns dynamically. For numeric columns, the mean is calculated, while for
character or factor columns, unique values are concatenated and separated by "|".
}
