% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateKUDs.R
\name{calculateKUDs}
\alias{calculateKUDs}
\title{Calculate animals' kernel utilization areas}
\usage{
calculateKUDs(
  data,
  bandwidth,
  grid,
  subset = NULL,
  id.groups = NULL,
  land.shape = NULL,
  id.col = getDefaults("id"),
  lon.col = getDefaults("lon"),
  lat.col = getDefaults("lat"),
  epsg.code = getDefaults("epsg")
)
}
\arguments{
\item{data}{A data frame with animal positions (COAs), containing 'longitude' and 'latitude' columns.}

\item{bandwidth}{Kernel bandwidth (h) - smoothing parameter used to calculate KUDs.}

\item{grid}{SpatialPixels object containing the grid over which animal KUDs are going to be estimated
(see grid argument in the \code{\link[adehabitatHR]{kernelUD}} function).}

\item{subset}{If defined, KUDs are calculated independently for each level of this variable.
If left NULL, single KUDs are calculated for the whole monitoring period.}

\item{id.groups}{Optional. A list containing ID groups, used to calculate stats independently
within each group.}

\item{land.shape}{Optional. A shapefile containing coastlines or landmasses. It can be supplied as
an 'sf' object or as an object of class 'SpatialPolygonsDataFrame' or 'SpatialPolygons'.
If the provided object is not of class 'sf', the function will attempt to
convert it to an 'sf' object for compatibility with subsequent spatial operations.}

\item{id.col}{Name of the column containing animal IDs. Defaults to 'ID'.}

\item{lon.col}{Name of the column containing longitude values. Defaults to 'lon'.}

\item{lat.col}{Name of the column containing latitude values. Defaults to 'lat'.}

\item{epsg.code}{The EPSG code (integer) representing the coordinate reference system (CRS) to be used
for projecting positions/layers.}
}
\value{
A list containing the estimated UDs, areas corresponding to
50\% and 95\% of occurrence probability, and a formatted data frame.
}
\description{
Pipeline to calculate animals' utilization areas using kernel
density estimation, based on \code{\link[adehabitatHR]{kernelUD}} function. It estimates
areas corresponding to 50\% and 95\% of occurrence probability (automatically subtracting
portions over land masses) and allows estimation of KUDs independently by group.
}
\seealso{
\code{\link[adehabitatHR]{kernelUD}}
}
