% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateLandDists.R
\name{calculateLandDists}
\alias{calculateLandDists}
\title{Calculate distance to nearest shore/land feature and other related metrics.}
\usage{
calculateLandDists(
  data,
  land.shape,
  mov.threshold = 0.5,
  id.col = "ID",
  lon.col = "lon",
  lat.col = "lat",
  dist.col = NULL
)
}
\arguments{
\item{data}{A data frame with animal detections/positions, containing longitude and latitude values (unprojected).}

\item{land.shape}{A shapefile containing coastlines or land surfaces.}

\item{mov.threshold}{Movement threshold (0 - 1). It establishes the proportion of a step length that must be
perpendicular to shore (away or approaching) in order for a movement to be classified as inshore or offshore
(e.g., with the default mov.threshold of 0.5: if the difference in distance to land between each two
consecutive positions corresponds to more than 50\% of the total step length, the movement
is classified either as inshore or offshore, according to its direction).}

\item{id.col}{Name of the column containing animal IDs Defaults to 'ID'.}

\item{lon.col}{Name of the column containing longitude values. Defaults to 'lon'.}

\item{lat.col}{Name of the column containing latitude values. Defaults to 'lat'.}

\item{dist.col}{Name of the column containing step lengths in meters (distance between each consecutive position).
If left null, the function only calculates distance to land and no additional metrics.}
}
\value{
A data frame containing distances to nearest shore and multiple related metrics, including
direction of movement (inshore/offshore/alongshore) and percentage of movement
perpendicular to shore (land_ratio).
}
\description{
Function to calculate distance to nearest land and additional related metrics,
including direction of movement (inshore/offshore/alongshore).
}
