% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateOverlap.R
\name{calculateOverlap}
\alias{calculateOverlap}
\title{Calculate pairwise spatio-temporal overlaps (association index)}
\usage{
calculateOverlap(
  table,
  id.groups = NULL,
  subset = NULL,
  metric = "simple-ratio",
  group.comparisons = "all",
  cores = 1
)
}
\arguments{
\item{table}{A data frame containing binned detections in the wide format
(time bin x individual matrix, with values corresponding to the receiver/station with the
highest number of detections), as returned by \code{\link{createWideTable}}.}

\item{id.groups}{Optional. A list containing ID groups, used to calculate stats independently
within each group, as well as comparing relationships between ids of different groups.}

\item{subset}{If defined, overlaps are calculated independently for each level of
this variable. This can either be a single column name (variable) or a vector
of column names, corresponding to variables contained in the table. In the case of
multiple columns, their interaction is used for grouping. If left NULL,
single overlap indices are calculated for the whole monitoring period.}

\item{metric}{One of "simple-ratio" or "half-weight".}

\item{group.comparisons}{Controls the type of comparisons to be run, when id.groups are defined.
One of "within", "between" or "all". Useful to discard comparisons between individuals
belonging to the same group or skip comparisons between different groups, when these are
not required (less computing time). Defaults to "all".}

\item{cores}{Number of CPU cores to use for the computations. Defaults to 1, which
means no parallel computing (single core).  If set to a value greater than 1,
the function will use parallel computing to speed up calculations.
Run \code{parallel::detectCores()} to check the number of available cores.}
}
\value{
A data frame containing the following columns:
\describe{
\item{id1}{ID of the first individual in the pair.}
\item{id2}{ID of the second individual in the pair.}
\item{overlap}{Pairwise overlap percentage between the two individuals.}
\item{co_occurrences}{Total number of time bins where both individuals were detected together.}
\item{shared_monit_days}{Number of days the two individuals were monitored simultaneously.}
\item{start}{Start date of the dyad shared monitoring period.}
\item{end}{End date of the dyad shared monitoring period.}
\item{mean_consec_overlap}{Average number of consecutive time bins where both individuals were detected together.}
\item{max_consec_overlap}{Maximum number of consecutive time bins where both individuals were detected together.}
}
If \code{subset} variable is defined, these results are nested within each of the grouping levels.
If \code{id.groups} is defined, the function will add the respective groups containing each one of the IDs.

A list containing similarity matrices with pairwise overlaps in \%,
time windows in days (period used to compute overlap for each pair),
the average number of time-bins spent by each pair consecutively in the same receiver,
the maximum number of time-bins spent by each pair consecutively in the same receiver and
the total number of co-occurrences, as well as the number of co-occurrences by location level.
If id.groups or a subset variable is defined, these results are nested within each of the grouping levels.
}
\description{
Estimates the extent of association between each pair of individuals,
assuming that joint space usage occurs whenever individuals overlap in space (receiver) and time (bin).
The analysis only considers the total shared periods of detection for a given pair, i.e., the time bins from the
latest release date to the earliest last detection between each pair. This truncation step ensures that the absence
of an individual in the receiver array is not due to transmitter failure or premature death. The pairwise overlap
index can be calculated using one of two metrics: the simple ratio association index
or the half-weight index. These indices range from 0\% (no overlap) to 100\% (complete overlap).
}
\details{
The two metrics for calculating association indices are:
\enumerate{
\item Simple Ratio Association Index:
This index provides a measure of the proportion of time two individuals spend together.
It is calculated as the ratio of the number of time bins where both individuals are
detected together to the total number of time bins where at least one of the individuals
is detected. This straightforward metric gives a clear indication of the extent
to which two individuals' space usage overlaps.
\item Half-Weight Index:
This index modifies the simple ratio by giving half-weight to the occurrences
where only one of the individuals is detected. It accounts for the fact that sometimes
individuals might be in the same area but one might not be detected due to various reasons.
This index is useful when detections are not perfectly reliable and helps mitigate
the impact of missed detections.
}

It is important to acknowledge that any fish pair detected simultaneously may
be anywhere from a few centimetres to hundreds of meters apart, depending on
the maximum detection ranges of the transmitters. Therefore, caution should
be taken when interpreting the results, especially when making inferences about
biotic or fish−habitat relationships.

Useful references:
\itemize{
\item Cairns, S. J., & Schwager, S. J. (1987). A comparison of association indices. Animal Behaviour, 35(5), 1454-1469.
\item Ginsberg, J. R., & Young, T. P. (1992). Measuring association between individuals or groups in behavioural studies. Animal Behaviour, 44(2), 377-379.
\item Farine, D. R., & Whitehead, H. (2015). Constructing, conducting and interpreting animal social network analysis. Journal of Animal Ecology, 84(5), 1144-1163.
\item Hoppitt, W. J., & Farine, D. R. (2018). Association indices for quantifying social relationships: how to deal with missing observations of individuals or groups. Animal Behaviour, 136, 227-238.
}
}
