% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateTracks.R
\name{calculateTracks}
\alias{calculateTracks}
\title{Estimate minimum traveled distances}
\usage{
calculateTracks(
  data,
  land.shape = NULL,
  epsg.code = getDefaults("epsg"),
  grid.resolution = 100,
  mov.directions = 16,
  id.col = getDefaults("ID"),
  lon.col = getDefaults("lon"),
  lat.col = getDefaults("lat"),
  cores = 1,
  verbose = TRUE
)
}
\arguments{
\item{data}{A data frame with animal positions, containing longitude and latitude values.}

\item{land.shape}{Optional. A shapefile containing coastlines or landmasses. It can be supplied as
an 'sf' object or as an object of class 'SpatialPolygonsDataFrame' or 'SpatialPolygons'.
If the provided object is not of class 'sf', the function will attempt to
convert it to an 'sf' object for compatibility with subsequent spatial operations.}

\item{epsg.code}{The EPSG code (integer) representing the coordinate reference system (CRS) to be used
for projecting the positions. If not specified, the function will attempt to use the CRS from the
provided land.shape (if available).}

\item{grid.resolution}{The grid cell size (in meters) used to estimate shortest paths. A higher
resolution leads to more precise path calculations but may increase computation time.}

\item{mov.directions}{Number of movement directions allowed for shortest path calculations, passed
to the \link[gdistance]{transition} function from the \strong{gdistance} package.}

\item{id.col}{Name of the column containing animal IDs. Defaults to 'ID'.}

\item{lon.col}{Name of the column containing longitude values. Defaults to 'lon'.}

\item{lat.col}{Name of the column containing latitude values. Defaults to 'lat'.}

\item{cores}{Number of CPU cores to use for the computations. Defaults to 1, which
means no parallel computing (single core).  If set to a value greater than 1,
the function will use parallel computing to speed up calculations. This parameter
is only relevant for estimating least-cost paths, specifically when a \code{land.shape} is provided.
Run \code{parallel::detectCores()} to check the number of available cores.}

\item{verbose}{Logical. Should the function output process information and display a progress bar?
Defaults to TRUE.}
}
\value{
A list with the following components:
\item{data}{A data frame containing the calculated (minimum) distances between consecutive animal positions.}
\item{trajectories}{A list of spatial objects representing the paths between positions.}
\item{transition_layer}{(If land.shape is provided) The transition cost layer used for shortest path calculations.}
}
\description{
This function calculates the shortest path in water between
consecutive positions for multiple individuals. If no land shapefile is provided,
the function defaults to calculating  linear (great-circle) distances. When a land
shapefile is provided, the function can utilize parallel computing to expedite
least-cost path estimation, depending on the number of CPU cores specified.
}
\note{
This function relies on the \strong{gdistance} and \strong{geosphere} packages for calculating
the shortest in-water paths and great-circle distances, respectively.
}
