% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctPositions.R
\name{correctPositions}
\alias{correctPositions}
\title{Relocate points on land to the nearest marine cell.}
\usage{
correctPositions(
  positions,
  layer,
  layer.type = "land",
  depth.margin = 0,
  lon.col = "lon",
  lat.col = "lat",
  plot = F
)
}
\arguments{
\item{positions}{A data frame containing longitude and latitude values
(unprojected lat-long coordinates), corresponding to the inferred animals' positions.}

\item{layer}{A raster containing land portions. Or alternatively a raster containing
elevation values (negative below water).}

\item{layer.type}{Either 'land', 'ocean' or 'bathy'. If type 'land' (binary) the raster is assumed
to contain NA values in water surfaces, if type 'ocean' (binary) the raster is assumed to
contain NA values in land surfaces, else if type 'bathy' the raster is assumed to
range between 0 (land) to max depth (positive or negative values).}

\item{depth.margin}{If set, shortest in-water paths will be calculated only
for cells with depths >= than this threshold. Only takes effect when raster
is of type bathy. Defaults to 0 (water vs land).}

\item{lon.col}{Name of the column containing longitude values. Defaults to 'lon'.}

\item{lat.col}{Name of the column containing latitude values. Defaults to 'lat'.}

\item{dist.margin}{If set, shortest in-water paths will be calculated only
for cells with depths <= than this threshold. Only takes effect when raster
is of type bathy. Disabled by default.}
}
\description{
Function to relocate positions on land to nearest marine cell
}
\details{
Either a coastline shapefile or a raster containing land surfaces or bathymetry values
can be provided.
}
