% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctPositions.R
\name{correctPositions}
\alias{correctPositions}
\title{Relocate points on land to the nearest marine cell.}
\usage{
correctPositions(
  positions,
  layer,
  layer.type = "land",
  depth.margin = 0,
  lon.col = getDefaults("lon"),
  lat.col = getDefaults("lat"),
  plot = F
)
}
\arguments{
\item{positions}{A data frame containing longitude and latitude values (unprojected lat-long coordinates),
corresponding to the inferred animals' positions.}

\item{layer}{A spatial object representing the geographic layer used to determine land and marine areas.
This can be either:\cr
\itemize{
\item A 'RasterLayer' containing binary values or bathymetric information (negative values for underwater).
\item A 'SpatialPolygons' object representing the coastline.
}}

\item{layer.type}{Either 'land', 'ocean' or 'bathy'. If type 'land' (binary) the raster is assumed
to contain NA values in water surfaces, if type 'ocean' (binary) the raster is assumed to
contain NA values in land surfaces, else if type 'bathy' the raster is assumed to
range between 0 (land) to max depth (positive or negative values).}

\item{depth.margin}{A numeric value. If set, shortest in-water paths will be calculated only
for cells with depths >= this threshold. Only takes effect when raster
is of type 'bathy'. Defaults to 0 (water vs land).}

\item{lon.col}{Name of the column containing longitude values. Defaults to 'lon'.}

\item{lat.col}{Name of the column containing latitude values. Defaults to 'lat'.}

\item{plot}{Logical, if TRUE, the function will plot the results. Defaults to FALSE.}
}
\value{
A data frame with corrected positions where points on land are relocated
to the nearest marine cell.
}
\description{
This function relocates positions on land to the nearest marine cell
using either a coastline shapefile or a raster containing land surfaces or bathymetry values.
}
