% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createWideTable.R
\name{createWideTable}
\alias{createWideTable}
\title{Create detections table in wide format}
\usage{
createWideTable(
  data,
  id.col = getDefaults("id"),
  timebin.col = getDefaults("timebin"),
  value.col,
  start.dates = NULL,
  end.dates = NULL,
  agg.fun = NULL,
  round.dates = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{data}{A data frame containing animal detections with corresponding time-bins.}

\item{id.col}{Name of the column containing animal IDs. Defaults to 'ID'.}

\item{timebin.col}{Name of the column containing time bins (in POSIXct format). Defaults to 'timebin'.}

\item{value.col}{The column name in the data frame whose values will be assigned
to each entry in the resulting wide-format table. If set to "detections" and no
such column exists in the dataset, the function will assume one detection per row
and create a corresponding column.}

\item{start.dates}{Optional. A POSIXct vector containing the start date of the monitoring period of each
animal (e.g., tag/release date of each individual). If left null, defaults to each individual's first detection.
If a single value is provided, it will be used for all IDs.}

\item{end.dates}{Optional. A POSIXct vector containing a cut-off date for each animal.
If left null, defaults to the individual's last detection.
If a single value is provided, it will be used for all IDs.}

\item{agg.fun}{A function to determine how values within each time bin are aggregated.
If left NULL, the aggregation method is automatically chosen based on the type of \code{value.col}:
for character or factor columns (e.g., 'station'), the most frequent value (mode) is used;
for numeric columns (e.g., 'detections'), the sum of values is used.}

\item{round.dates}{Logical. If TRUE, the start and end dates are rounded to the nearest day:
the earliest start date is floored to the beginning of the day, and the latest end date is
rounded up to the end of the day. This ensures that the time bins cover whole days. Defaults to FALSE.}

\item{verbose}{Output ties info to console? Defaults to TRUE.}
}
\description{
Returns a data frame object containing binned detections in the wide format
(time bin x individual matrix, with values corresponding either to the receiver detections
or the number of detections).
}
