% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createWideTable.R
\name{createWideTable}
\alias{createWideTable}
\title{Create detections table in wide format.}
\usage{
createWideTable(
  data,
  id.col = "ID",
  timebin.col = "timebin",
  value.col,
  start.dates = NULL,
  end.dates = NULL,
  agg.fun = NULL,
  round.dates = F,
  verbose = T
)
}
\arguments{
\item{data}{A data frame containing animal detections with corresponding time-bins.}

\item{id.col}{Name of the column containing animal IDs. Defaults to 'ID'.}

\item{timebin.col}{Name of the column containing time bins (in POSIXct format). Defaults to 'timebin'.}

\item{value.col}{The values to be assigned to each entry, should match a column
in the supplied data (except if set to "detections"). If set to "detections" and no detections
column is found in the dataset, the function automatically assumes one detection per row.}

\item{start.dates}{Optional. A POSIXct vector containing the start date of the monitoring period of each
animal (e.g.tag/release date of each individual). If left null, defaults to each individual's first detection.
If a single value is provided, it will be used to all IDs.}

\item{end.dates}{Optional. A POSIXct vector containing the tag/release date
of each animal.If left null, defaults to the individual's last detections.
If a single value is provided, it will be used to all IDs.}

\item{agg.fun}{Function to assign the final value across each time bin.
By default (if left null), it automatically adjusts based on the class type of the value.col.
If this column is of type character or factor (e.g., 'station' or 'habitat'),
the value is assigned based on the most frequently occurring observation within each timebin (i.e., mode).
If it is numeric (e.g., number of detections), the assigned value will correspond to
the sum of all observations within each timebin.}

\item{round.dates}{Boolean to indicate if start and end dates should be rounded
(floor of earliest tagging and ceiling of last detection).}

\item{verbose}{Output ties info to console? Defaults to TRUE.}
}
\description{
Returns a data frame object containing binned detections in the wide format
(time bin x individual matrix, with values corresponding either to the receiver > detections
or the number of detections).
}
