% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateArguments.R
\name{.validateArguments}
\alias{.validateArguments}
\title{Validate Arguments}
\usage{
.validateArguments()
}
\arguments{
\item{...}{Arguments to validate. This function dynamically retrieves and checks the arguments
passed to the calling function. The first argument should be a data frame containing the telemetry data.}
}
\value{
A list containing potentially modified versions of \code{data}, \code{tagging.dates}, and \code{tag.durations}.
If validation fails, the function will stop and return an error message.
}
\description{
This internal function performs a comprehensive set of validation checks on the arguments
provided to functions within the \code{moby} package. It ensures that the data and various parameters
are in the correct format and meet the expected criteria, issuing errors and warnings as necessary.
The function tries to issue as many errors and warnings at the same time as possible, to speed up the
completion of required changes and reduce the time it takes to get the function call right.
}
\details{
The \code{validateArguments} function checks for the following:
\describe{
\item{data}{Ensures the first argument is a data frame.}
\item{id.col}{Verifies the ID column exists and converts it to a factor if necessary.}
\item{datetime.col}{Ensures the datetime column exists and is in POSIXct format.}
\item{timebin.col}{Checks the time-bin column for existence and correct format.}
\item{lon.col}{Ensures the longitude column exists and is numeric.}
\item{lat.col}{Ensures the latitude column exists and is numeric.}
\item{station.col}{Checks for the existence of the station column.}
\item{split.by}{Validates the column(s) used to split the data.}
\item{color.by}{Verifies the color-by column exists and contains no missing values; converts it to a factor if needed.}
\item{tagging.dates}{Checks tagging dates are in POSIXct format and match the number of unique IDs if required.}
\item{tag.durations}{Ensures tag durations are specified and correctly matched with unique IDs.}
\item{id.groups}{Validates that ID groups are unique and present in the data.}
\item{land.shape}{Checks if the provided spatial object can be converted to class \code{sf} if necessary.}
\item{color.pal}{Ensures the color palette is either a character vector or a function.}
\item{diel.lines}{Validates diel line values (0, 2, or 4).}
\item{polygons}{Checks if the polygons parameter is set to a valid value ('diel', 'season', or FALSE).}
\item{background.color}{Ensures the background color is a valid hex code or named color.}
\item{style}{Verifies the style is either 'raster' or 'points'.}
}
}
\note{
This function is intended for internal use within the \code{moby} package. It helps streamline
argument validation across multiple functions, ensuring consistency and reducing redundancy.
}
\keyword{internal}
