% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterDetections.R
\name{filterDetections}
\alias{filterDetections}
\title{Filter Detections}
\usage{
filterDetections(
  data,
  tagging.dates = getDefaults("tagging.dates"),
  cutoff.dates = NULL,
  id.col = getDefaults("id"),
  datetime.col = getDefaults("datetime"),
  lon.col = getDefaults("lon"),
  lat.col = getDefaults("lat"),
  min.detections = 0,
  min.days = 0,
  hours.threshold = 24,
  max.speed = NULL,
  speed.unit = "m/s",
  acoustic.range = 600,
  land.shape = NULL,
  epsg.code = getDefaults("epsg"),
  ...
)
}
\arguments{
\item{data}{A data frame containing raw animal detections.}

\item{tagging.dates}{A POSIXct vector containing the tag/release date of each animal.
The length of this vector should match the number of unique animal IDs.
Alternatively, if a single value is provided, it will be applied to all IDs.}

\item{cutoff.dates}{Optional. A POSIXct vector containing the estimated expiration dates of the tags
or any other cut-off date beyond which detections should be discarded. The length of this vector
should match the number of unique animal IDs. Alternatively, if a single value is provided, it will be applied to all IDs.}

\item{id.col}{Name of the column containing animal IDs. Defaults to 'ID'.}

\item{datetime.col}{Name of the column containing datetimes in POSIXct format. Defaults to 'datetime'.}

\item{lon.col}{Name of the column containing longitude values (unprojected). Defaults to 'lon'.}

\item{lat.col}{Name of the column containing latitude values (unprojected). Defaults to 'lat'.}

\item{min.detections}{Optional. Discard individuals with fewer than this number of detections.}

\item{min.days}{Optional. Minimum number of days an individual must be detected for it to be included;
individuals with fewer days of detections will be discarded.}

\item{hours.threshold}{Discard single detections occurring alone within a predefined time interval.}

\item{max.speed}{Maximum allowed speed between detections. If null, no speed-based checks are performed.}

\item{speed.unit}{Units of supplied maximum speed. Either 'm/s' or 'km/h'.}

\item{acoustic.range}{Maximum assumed detection range of the acoustic receivers (in metres).
Used in the speed filter to account for uncertainties in animal positions within the radius of the receivers,
when estimating minimum consecutive distances.}

\item{land.shape}{Optional. A  shape file containing coastlines.}

\item{epsg.code}{Coordinate reference system used to project positions (class 'CRS').
If not supplied, CRS is assumed to be the same as in land.shape.}

\item{...}{Further arguments passed to the \code{\link{calculateDistances}} function
(e.g., grid.resolution and mov.directions).}
}
\value{
A list containing both the filtered data and the rejected detections,
as well as a detailed summary of the filtered out data by individual.
}
\description{
This function filters out spurious animal detections based on
multiple criteria to ensure data accuracy. It helps users identify and remove
false positives and discard individuals with few or no valid detections.
It implements six distinct criteria:
\enumerate{
\item detections occurring before the animal's tagging date
\item detections occurring after an optional cut-off date
\item isolated detections within a specified time interval
\item detections implying unrealistic movement speeds
\item the total number of detections
\item the total number of logged days
}

Unlike other methods, \code{filterDetections} accounts for land topographies and
complex coastlines when calculating animal speeds if a
\code{land.shape} (shape file containing coastlines) is provided,
leading to more accurate distance estimates.
}
