% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterDetections.R
\name{filterDetections}
\alias{filterDetections}
\title{Filter Detections}
\usage{
filterDetections(
  data,
  tagging.dates = getDefaults("tagging.dates"),
  cutoff.dates = NULL,
  id.col = getDefaults("ID"),
  datetime.col = getDefaults("datetime"),
  lon.col = getDefaults("lon"),
  lat.col = getDefaults("lat"),
  min.detections = 0,
  min.days = 0,
  hours.threshold = 24,
  max.speed = NULL,
  speed.unit = "m/s",
  max.iterations = Inf,
  acoustic.range = 600,
  land.shape = NULL,
  epsg.code = getDefaults("epsg"),
  ...
)
}
\arguments{
\item{data}{A data frame containing raw animal detections.}

\item{tagging.dates}{A POSIXct vector specifying the tagging or release dates for each animal.
This parameter must be either:
\itemize{
\item A single POSIXct value, which will be applied to all unique animal IDs; or
\item A named POSIXct vector, where the names correspond to the animal IDs in the \code{id.col} column.
If multiple tagging dates are provided, the vector must include all IDs and will be reordered to align with the levels of \code{id.col}.
}}

\item{cutoff.dates}{Optional. A POSIXct vector containing the estimated expiration dates of the tags
or any other cut-off date beyond which detections should be discarded This parameter must be either:
\itemize{
\item A single POSIXct value, which will be applied to all unique animal IDs; or
\item A named POSIXct vector, where the names correspond to the animal IDs in the \code{id.col} column.
If multiple cut-off dates are provided, the vector must include all IDs and will be reordered to align with the levels of \code{id.col}.
}}

\item{id.col}{Name of the column containing animal IDs. Defaults to 'ID'.}

\item{datetime.col}{Name of the column containing datetimes in POSIXct format. Defaults to 'datetime'.}

\item{lon.col}{Name of the column containing longitude values. Defaults to 'lon'. Coordinates can be supplied
either in geographic (unprojected) or projected format. In both cases, the appropriate projected CRS
should be provided via the \code{epsg.code} argument.}

\item{lat.col}{Name of the column containing latitude values. Defaults to 'lat'. Coordinates can be supplied
either in geographic (unprojected) or projected format. In both cases, the appropriate projected CRS
should be provided via the \code{epsg.code} argument.}

\item{min.detections}{Optional. Discard individuals with fewer than this number of detections.}

\item{min.days}{Optional. Minimum number of days an individual must be detected for it to be included;
individuals with fewer days of detections will be discarded.}

\item{hours.threshold}{Discard single detections occurring alone within a predefined time interval.}

\item{max.speed}{Maximum allowed speed between detections. If null, no speed-based checks are performed.}

\item{speed.unit}{Units of supplied maximum speed. Either 'm/s' or 'km/h'.}

\item{max.iterations}{An integer specifying the maximum number of iterations
for the speed filtering process. The filtering loop will stop early if this number
of iterations is reached, even if detections are still being removed. The default is
\code{Inf} (unlimited iterations).}

\item{acoustic.range}{Maximum assumed detection range of the acoustic receivers (in metres).
Used in the speed filter to account for uncertainties in animal positions within the radius of the receivers,
when estimating minimum consecutive distances.}

\item{land.shape}{Optional. A shapefile containing coastlines or landmasses. It can be supplied as
an 'sf' object or as an object of class 'SpatialPolygonsDataFrame' or 'SpatialPolygons'.
If the provided object is not of class 'sf', the function will attempt to
convert it to an 'sf' object for compatibility with subsequent spatial operations.
If not supplied, the function will calculate linear (great-circle) distances instead of shortest in-water
distances. Linear distances are much faster to compute but may be less realistic or accurate in cases
where positions are separated by landmasses or other spatial obstacles. This parameter
is passed to the \code{\link{calculateTracks}} function and is only used when
a \code{max.speed} is defined.}

\item{epsg.code}{Coordinate reference system used to project positions (class 'CRS').
If not supplied, CRS is assumed to be the same as in land.shape.}

\item{...}{Further arguments passed to the \code{\link{calculateTracks}} function
(e.g., \code{grid.resolution}, \code{mov.directions} and \code{cores}).}
}
\value{
A list containing both the filtered data and the rejected detections,
as well as a detailed summary of the filtered out data by individual.
}
\description{
This function filters out spurious animal detections based on
multiple criteria to ensure data accuracy. It helps users identify and remove
false positives and discard individuals with few or no valid detections.
It implements six distinct criteria:
\enumerate{
\item detections occurring before the animal's tagging date
\item detections occurring after an optional cut-off date
\item isolated detections within a specified time interval
\item detections implying unrealistic movement speeds
\item the total number of detections
\item the total number of logged days
}

For speed filtering, the function iteratively removes detections that exceed the
maximum speed threshold, recalculating distances until no violations remain or the
maximum iteration limit is reached. When an unrealistic speed is detected, a hierarchical
clustering approach is applied to the spatial coordinates of neighboring detections.
This helps identify the outlier detection, likely responsible for triggering the overspeed
threshold, by considering its context within surrounding detections.

Additionally, \code{filterDetections} accounts for land topographies and complex
coastlines when calculating animal speeds if a \code{land.shape} (shapefile containing
coastlines or landmasses) is provided, leading to more accurate distance estimates.
}
