% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterDetections.R
\name{filterDetections}
\alias{filterDetections}
\title{Filter data}
\usage{
filterDetections(
  data,
  tagging.dates,
  id.col = "ID",
  min.detections = 0,
  min.days = 0,
  hours.threshold = 24,
  max.speed = NULL,
  speed.unit = "m/s",
  acoustic.range = 600,
  land.shape = NULL,
  epsg.code = NULL,
  ...
)
}
\arguments{
\item{data}{A data frame containing raw animal detections.}

\item{tagging.dates}{A POSIXct vector containing the tag/release date of each animal.}

\item{id.col}{Name of the column containing animal identifications. Defaults to "ID".}

\item{min.detections}{Optional. Discard individuals with less than this number of detections.}

\item{min.days}{Optional. Minimum number of days an individual must be detected for it to be included;
individuals with fewer detections will be discarded.}

\item{hours.threshold}{Discard single detections occurring alone within a predefined time interval.}

\item{max.speed}{Maximum allowed speed between detections. If null, no speed-based checks are performed.}

\item{speed.unit}{Units of supplied maximum speed. Either 'm/s' or 'km/h'.}

\item{acoustic.range}{Maximum assumed detection range of the acoustic receivers (in metres).
Used in the speed filter to account for uncertainties in animal positions within the radius of the receivers.}

\item{...}{Further arguments passed to the \code{\link{calculateDistances}} function
(e.g. land.shape, epsg.code, grid.resolution and mov.directions).}
}
\value{
A list containing both the filtered data and the rejected detections,
as well as a detailed summary of the filtered out data by individual.
}
\description{
Function to remove spurious detections, based on tagging date and/or
isolated occurrences.
}
