% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDielPhase.R
\name{getDielPhase}
\alias{getDielPhase}
\title{Estimate diel phase}
\usage{
getDielPhase(datetime, coordinates, phases = 2, solar.depth = 18)
}
\arguments{
\item{datetime}{A POSIXct object containing the respective datetimes or time-bins.}

\item{coordinates}{A SpatialPoints or matrix object containing longitude and
latitude coordinates (in that order) at which to estimate sunrise and sunset times.}

\item{phases}{Number of diel phases to return.}

\item{solar.depth}{Angle of the sun below the horizon (in degrees). Passed the
solarDep argument in \code{\link[suntools]{crepuscule}} function.}
}
\value{
Factor with diel phase
}
\description{
Function to retrieve diel phase (e.g. day/night) in which detections occurred.
Number of retrieved levels can be set by the 'phases' argument:
• phases=2:  day | night
• phases=3:  day| crepuscule | night
• phases=4:  dawn | day | dusk | night
Crepuscular periods can be defined based on the different solar elevation angles:
• solar.depth=6: civil twilight
• solar.depth=12: nautical twilight
• solar.depth=18: astronomical twilight
}
