% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReprodPeriod.R
\name{getReprodPeriod}
\alias{getReprodPeriod}
\title{Assign spawning periods}
\usage{
getReprodPeriod(date, spawning.start, spawning.end, format = "\%m")
}
\arguments{
\item{date}{A POSIXct object containing the respective datetimes or time-bins.}

\item{spawning.start}{Start of the spawning season. Can be supplied as
a POSIXct object or as month.}

\item{spawning.end}{End of the spawning season. Can be supplied as
a POSIXct object or as month.}

\item{format}{Character string giving the date-time format of the supplied
spawning.start and spawning.end.}
}
\value{
Factor with reproductive season (resting vs spawning).
}
\description{
Assign spawning vs resting periods based on given start and end interval.
}
