% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSeason.R
\name{getSeason}
\alias{getSeason}
\title{Estimate annual season.}
\usage{
getSeason(date, hemisphere = "Northern")
}
\arguments{
\item{date}{A POSIXct object containing the respective datetimes or time-bins.}

\item{hemisphere}{Earth hemisphere for which to calculate seasons.}
}
\value{
Factor with season
}
\description{
In temperate and sub-polar regions, 4 seasons based on the Gregorian calendar
(meteorological definition) are generally recognized:
Northern Hemisphere:
• Spring: March - May
• Summer: June - August
• Autumn: September - November
• Winter: December - February

Southern Hemisphere
• Spring: September - November
• Summer: December - February
• March - May
• Winter: June - August
}
