% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolateDistances.R
\name{interpolateDistances}
\alias{interpolateDistances}
\title{Distance interpolation}
\usage{
interpolateDistances(
  data,
  id.col = getDefaults("ID"),
  dist.col = "dist_m",
  keep.intermediate = FALSE
)
}
\arguments{
\item{data}{Output \code{\link{calculateTracks}} function: a data frame
with animal IDS, positions and distances.}

\item{id.col}{Name of the column containing animal IDs. Defaults to 'ID'.}

\item{dist.col}{Name of the column containing distances. Defaults to 'dist_m',
the output of \code{\link{calculateTracks}}.}

\item{keep.intermediate}{Boolean indicating if intermediate distances (assigned
to time-bins without detections) should be kept or discarded. Defaults to false.}
}
\value{
Original data frame plus missing time-bins (with diluted distances)
}
\description{
Function to interpolate (average) distances across all timebins.
If an animal goes undetected for large periods of time, large distance spikes
may occur after running the \code{\link{calculateTracks}} function.
Hence, to avoid large biases when averaging distances per hour or day, this
function 'dilutes' traveled distances by the total number of time-bins
passed between each two consecutive positions.
}
