% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movementTable.R
\name{movementTable}
\alias{movementTable}
\title{Create movement stats table}
\usage{
movementTable(
  data,
  kud.results,
  land.shape,
  epsg.code = NULL,
  transition.layer,
  id.groups = NULL,
  id.col = "ID",
  lon.col = "lon",
  lat.col = "lat",
  dist.col = "dist_m",
  discard.missing = T
)
}
\arguments{
\item{data}{A data frame containing binned animal detections and distances traveled,
as returned by \code{\link{calculateDistances}}.}

\item{kud.results}{Output of \code{\link{calculateKUDs}}.}

\item{land.shape}{A projected shape file containing coastlines.}

\item{epsg.code}{Coordinate reference system used to project positions (class 'CRS').
If not supplied, CRS is assumed to be the same as in land.shape.}

\item{transition.layer}{Transition layer used to calculate shortest in-water paths, as
retrieved by the \code{\link{calculateDistances}} function.}

\item{id.groups}{Optional. A list containing ID groups, used to
visually aggregate animals belonging to the same class (e.g. different species).}

\item{id.col}{Name of the column containing animal IDs Defaults to 'ID'.}

\item{lon.col}{Name of the column containing longitude values (unprojected). Defaults to 'lon'.}

\item{lat.col}{Name of the column containing latitude values (unprojected). Defaults to 'lat'.}

\item{dist.col}{Name of the column containing distance values (in meters). Defaults to 'dist_m'.}

\item{discard.missing}{If true, only individuals with detections are included.}
}
\description{
Creates a table containing  movement metrics for each animal
(total distances traveled, rate of movements, KUDs, etc.).
}
