% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movementTable.R
\name{movementTable}
\alias{movementTable}
\title{Create movement stats table}
\usage{
movementTable(
  data,
  kud.results,
  land.shape,
  epsg.code = getDefaults("epsg"),
  id.groups = NULL,
  id.col = getDefaults("ID"),
  timebin.col = getDefaults("timebin"),
  lon.col = getDefaults("lon"),
  lat.col = getDefaults("lat"),
  dist.col = "dist_m",
  discard.missing = TRUE,
  ...
)
}
\arguments{
\item{data}{A data frame containing binned animal detections and distances traveled,
as returned by \code{\link{calculateTracks}}.}

\item{kud.results}{Output of \code{\link{calculateKUDs}}.}

\item{land.shape}{A projected shape file containing coastlines.}

\item{epsg.code}{Coordinate reference system used to project positions (class 'CRS').
If not supplied, CRS is assumed to be the same as in land.shape.}

\item{id.groups}{Optional. A list containing ID groups, used to
visually aggregate animals belonging to the same class (e.g. different species).}

\item{id.col}{Name of the column containing animal IDs. Defaults to 'ID'.}

\item{timebin.col}{Name of the column containing time bins (in POSIXct format). Defaults to 'timebin'.}

\item{lon.col}{Name of the column containing longitude values. Defaults to 'lon'. Coordinates can be supplied
either in geographic (unprojected) or projected format. In both cases, the appropriate projected CRS
should be provided via the \code{epsg.code} argument.}

\item{lat.col}{Name of the column containing latitude values. Defaults to 'lat'. Coordinates can be supplied
either in geographic (unprojected) or projected format. In both cases, the appropriate projected CRS
should be provided via the \code{epsg.code} argument.}

\item{dist.col}{Name of the column containing distance values (in meters). Defaults to 'dist_m'.}

\item{discard.missing}{If true, only individuals with detections are included.}

\item{...}{Further arguments passed to \code{\link{calculateTracks}} function,
in order to calculate distances between the first and last recorded detection, for each individual.}
}
\description{
Creates a table containing  movement metrics for each animal
(total distances traveled, rate of movements, KUDs, etc.).
}
