% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotChronogram.R
\name{plotChronogram}
\alias{plotChronogram}
\title{Generate chronograms (heatmaps)}
\usage{
plotChronogram(
  data,
  tagging.dates = getDefaults("tagging.dates"),
  variables = "detections",
  split.by = NULL,
  style = "points",
  id.col = getDefaults("id"),
  timebin.col = getDefaults("timebin"),
  station.col = getDefaults("station"),
  color.by = NULL,
  color.pal = NULL,
  date.format = "\%d/\%b",
  date.interval = 4,
  date.start = 1,
  diel.lines = 4,
  sunriset.coords,
  solar.depth = 18,
  polygons = F,
  lunar.info = F,
  background.col = "white",
  grid = TRUE,
  grid.color = "white",
  highlight.isolated = FALSE,
  pt.cex = c(0.5, 3),
  uniformize.scale = FALSE,
  uniformize.dates = TRUE,
  cex.axis = 1,
  cex.legend = 0.8,
  cex.moon = 2.5,
  legend.intersp = 1.2,
  legend.cols = NULL,
  cols = NULL,
  ...
)
}
\arguments{
\item{data}{A data frame containing binned animal detections.}

\item{tagging.dates}{A POSIXct vector containing the tag/release date of each animal.
The length of this vector should match the number of unique animal IDs.
Alternatively, if a single value is provided, it will be applied to all IDs.}

\item{variables}{The type(s) of metric to plot. Accepted types: "detections", "individuals" and "co-occurrences".}

\item{split.by}{Optional. If defined, plots are generated individually for each level
of this variable (e.g. species, ontogeny or habitat).}

\item{style}{Style of the plot. Either "raster" (nº represented by color) or "points" (nº represented by size).}

\item{id.col}{Name of the column containing animal IDs. Defaults to 'ID'.}

\item{timebin.col}{Name of the column containing time bins (in POSIXct format). Defaults to 'timebin'.}

\item{station.col}{Name of the column containing station/receiver IDs. Defaults to 'station'.}

\item{color.by}{Variable defining the color group of the plotted metric, when style = "points".
Can be used for example to display detections by receiver, animal trait or temporal category.}

\item{color.pal}{Color palette for the level plot.}

\item{date.format}{Date-time format (as used in \code{\link[base]{strptime}}),
defining the x-axis labels. Defaults to month ("\%b").}

\item{date.interval}{Number defining the interval between each
displayed date (x-axis label). Defaults to 4.}

\item{date.start}{Integer defining the first displayed date (can be used in combination
with 'date.interval" to better control the x-axis labels). Defaults to 1.}

\item{diel.lines}{Number indicating the number of diel phase lines (boundaries)
to display. Either 0 (no lines), 2 (corresponding to sunrise and sunset) or 4
(depicting dawn, sunrise, sunset, dusk). Defaults to 4.}

\item{sunriset.coords}{A SpatialPoints or matrix object containing longitude and
latitude coordinates (in that order) at which to estimate sunrise and sunset times.}

\item{solar.depth}{Angle of the sun below the horizon (in degrees). Passed the
solarDep argument in \code{\link[maptools]{crepuscule}} function.}

\item{polygons}{Either "diel", "season". If style is set to points, can be used to highlight
diel or seasonal phases using color-coded polygons. Defaults to F (no polygons).}

\item{lunar.info}{Include top mural containing lunar illumination?}

\item{background.col}{If style is set to points, defines the background color of the plot.}

\item{grid}{If true, a grid is plotted (horizontal and vertical guides across the entire plot).}

\item{grid.color}{Color of the grid lines, if 'grid' is set to true.}

\item{highlight.isolated}{If a 'color.by' variable is defined and style="points, points are ordered
and plotted according to their "density", with isolated levels being brought forward
to prevent them from being hidden behind denser point clouds.}

\item{pt.cex}{Defines the min and max point radius when style="point. Defaults to c(0,3).}

\item{uniformize.scale}{Use the same color scale for all plots when 'split.by' is defined.}

\item{uniformize.dates}{Use the same date range for all plots when 'split.by' is defined.}

\item{cex.axis}{Determines the size of the text labels on the axes. Defaults to 1.}

\item{cex.legend}{Determines the size of the color legend. Defaults to 0.8.}

\item{cex.moon}{Determines the size of the moon phase symbols, when lunar.info is set to TRUE. Defaults to 2.5.}

\item{legend.intersp}{Vertical distances between legend elements
(in lines of text shared above/below each legend entry). Defaults to 1.2.}

\item{legend.cols}{Integer. The number of columns in which to set the legend items.
If NULL, it is set automatically based on the number of levels. Defaults to NULL.}

\item{cols}{Number of columns in the final panel (passed to the mfrow argument).}

\item{...}{Further arguments passed to the \code{\link[graphics]{image}} function (if style="raster"),
or to the \code{\link[graphics]{plot}} function (if style="points").}
}
\description{
Function to represent a given metric (nº detections, presences, individuals or co-occurrences)
per time bin. It generates 2-dimensional plots (hour x date) highlighting the chosen variable
either through color-coded cells or size-variable points, and illustrating the
variation in diel phases' hours across the study duration (annual variation of daylight time).
}
