% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotContours.R
\name{plotContours}
\alias{plotContours}
\title{Generate contour plot}
\usage{
plotContours(
  data,
  variables,
  var.titles = NULL,
  plot.title = NULL,
  split.by = NULL,
  aggregate.fun = function(x) mean(x, na.rm = T),
  id.col = getDefaults("id"),
  datetime.col = getDefaults("datetime"),
  color.pal = NULL,
  diel.lines = 4,
  diel.lines.col = "black",
  sunriset.coords = NULL,
  solar.depth = 18,
  cex.main = 1.1,
  cex.lab = 1,
  cex.axis = 0.8,
  grid = TRUE,
  invert.scale = FALSE,
  uniformize.scale = FALSE,
  legend.xpos = c(0.89, 0.92),
  legend.ypos = c(0.15, 0.85),
  tz = "UTC",
  cols = 1,
  disable.par = FALSE,
  ...
)
}
\arguments{
\item{data}{A data frame containing the variable(s) to plot.}

\item{variables}{Variables to plot. Should correspond columns in the data.
Can be either a vector or a single variable.}

\item{var.titles}{Display name of the variables (e.g. "Depth (m)"). If NULL, defaults to supplied variable names.}

\item{plot.title}{Optional. Main title displayed on top of the plots/panel.}

\item{split.by}{Optional. If defined, plots are generated individually for each level (or combination of levels)
of this variable(s) (e.g. species, ontogeny, sex, habitat, etc).}

\item{aggregate.fun}{Function used to aggregate values by hour and month. Defaults to mean.}

\item{id.col}{Name of the column containing animal IDs. Defaults to 'ID'.}

\item{datetime.col}{Name of the column containing datetimes in POSICX format. Defaults to 'datetime'.}

\item{color.pal}{Color palette for the contour plot. If NULL, defaults to the viridis color palette.}

\item{diel.lines}{Number indicating the number of diel phase lines (boundaries)
to display. Either 0 (no lines), 2 (corresponding to sunrise and sunset) or 4
(depicting dawn, sunrise, sunset, dusk). Defaults to 4.}

\item{diel.lines.col}{Color of the diel lines. Defaults to black.}

\item{sunriset.coords}{A SpatialPoints or matrix object containing longitude and
latitude coordinates (in that order) at which to estimate sunrise and sunset times
(if \code{diel.lines} > 0).}

\item{solar.depth}{Angle of the sun below the horizon (in degrees). Passed to the
\code{solarDep} argument in \code{\link[suntools]{crepuscule}} function. Defaults to 18 (astronomical twilight).}

\item{cex.main}{Determines the size of the title(s). Defaults to 1.1.}

\item{cex.lab}{Determines the size of the y-axis and y-axis labels. Defaults to 1.}

\item{cex.axis}{Determines the size of the text labels on the axes. Defaults to 0.8.}

\item{grid}{Add horizontal and vertical guiding lines? Defaults to True.}

\item{invert.scale}{Invert color scale legend? Defaults to False.}

\item{uniformize.scale}{Use the same color scale for all plots when 'split.by' is defined.}

\item{legend.xpos}{Relative position of left and right edge of color bar on first axis (0-1).
Defaults to c(0.89, 0.92).}

\item{legend.ypos}{Relative position of left and right edge of color bar on second axis (0-1).
Defaults to c(0.15, 0.85).}

\item{tz}{A character string specifying the time zone to be used for the conversion. Defaults to "UTC".}

\item{cols}{Number of columns in the final panel (passed to the mfrow argument). Only used when
by.id is set to true or the split.by argument is defined.}

\item{disable.par}{Boolean. Disables internal par settings, so that multi-figure/panel layouts can be
generated if required. If set to True, 'cols' argument will have no effect in the output. Defaults to False.}

\item{...}{Further arguments passed to the \code{\link{filled.contour}} function}
}
\description{
Function to generate a contour plot (color-coded 2-dimensional plots)
of a given variable, by hour and month, either for all individuals combined (average) or
independently for each ID. It also plots lines illustrating the variation in
diel phases' hours across the study duration (annual variation of daylight time).
}
