% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCWTs.R
\name{plotCWTs}
\alias{plotCWTs}
\title{Continuous wavelet transform (CWT)
Decompose a signal into its constituent frequency components over time.}
\usage{
plotCWTs(
  data,
  variable,
  id.col = "ID",
  period.range = c(3, 48),
  axis.periods = c(6, 12, 16, 24, 48),
  color.pal = NULL,
  date.format = "\%d/\%b",
  date.interval = 4,
  date.start = 1,
  min.days = NULL,
  detrend = F,
  cex.main = 1.2,
  cex.lab = 1.1,
  cex.axis = 1,
  legend.xpos = c(0.9, 0.915),
  legend.ypos = c(0.15, 0.85),
  cols = 2,
  same.scale = F,
  id.groups = NULL,
  ...
)
}
\arguments{
\item{data}{A data frame containing binned animal detections (must contain a "timebin" column).}

\item{variable}{Name of the column containing the numeric variable to be analyzed.}

\item{id.col}{Name of the column containing animal IDs. Defaults to 'ID'.}

\item{period.range}{The range of period scales (y-axis limits) to be considered, specified in hours.}

\item{axis.periods}{Periods (in hours) to include/highlight in the y-axis.}

\item{color.pal}{Color palette. Defaults to \code{\link[pals]{jet}}}

\item{date.format}{Date-time format (as used in \code{\link[base]{strptime}}),
defining the x-axis labels. Defaults to month ("\%d/\%b").}

\item{date.interval}{Number defining the interval between each
displayed date (x-axis label). Defaults to 4.}

\item{date.start}{Integer defining the first displayed date (can be used in combination
with 'date.interval" to better control the x-axis labels). Defaults to 1.}

\item{min.days}{Discard individuals that were detected in less than x days.}

\item{detrend}{Detrend time series using differences (\code{\link[base]{diff}})
rather than the actual values. Defaults to False.}

\item{cex.main}{Determines the size of the title(s). Defaults to 1.2.}

\item{cex.lab}{Determines the size of the y-axis and y-axis labels. Defaults to 1.1.}

\item{cex.axis}{Determines the size of the text labels on the axes. Defaults to 1.}

\item{legend.xpos}{Relative position of left and right edge of color bar on first axis (0-1).
Passed to \code{\link[shape]{colorlegend}} function. Defaults to c(0.90, 0.915).}

\item{legend.ypos}{Relative position of left and right edge of color bar on second axis (0-1).
Passed to \code{\link[shape]{colorlegend}} function. Defaults to c(0.15, 0.85).}

\item{cols}{Number of columns in the final panel (passed to the mfrow argument).}

\item{same.scale}{Forces same spectral scale (zlims) across all plots,
allowing for density comparison between individuals.}

\item{id.groups}{Optional. A list containing ID groups, used to
visually aggregate animals belonging to the same class (e.g. different species).}

\item{...}{Further arguments passed to the \code{\link[wavScalogram]{cwt_wst}} function
(e.g. wname, border_effects, waverad).}
}
\description{
Function that analyzes and plots periodic patterns in a time series using a
Continuous Wavelet Transform (CWT) framework. CWT analysis provides an alternative method
to the Fast Fourier Transform (FFT) or other time-frequency decomposition techniques,
enabling the examination of periodicities over time. This function serves mostly as a
wrapper for the \code{\link[wavScalogram]{cwt_wst}} function.
}
