% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDetections.R
\name{plotDetections}
\alias{plotDetections}
\title{Detection plots}
\usage{
plotDetections(
  data,
  id.col = getDefaults("id"),
  datetime.col = getDefaults("datetime"),
  tagging.dates = getDefaults("tagging.dates"),
  tag.durations = NULL,
  id.groups = NULL,
  discard.missing = T,
  color.by = NULL,
  color.pal = NULL,
  date.format = "\%b/\%y",
  date.interval = 4,
  date.start = 1,
  sunriset.coords,
  diel.lines = 2,
  background.color = "gray96",
  highlight.isolated = T,
  pt.cex = 1.6,
  grid = F,
  grid.color = "white",
  cex.title = 2.2,
  cex.lab = 1.8,
  cex.axis = 1.4,
  cex.legend = 1.6,
  cols = 2,
  legend.cols = 3
)
}
\arguments{
\item{data}{A data frame containing animal detections.}

\item{id.col}{Name of the column containing animal IDs. Defaults to 'ID'.}

\item{datetime.col}{Name of the column containing datetimes in POSIXct format. Defaults to 'datetime'.}

\item{tagging.dates}{A POSIXct vector containing the tag/release date of each animal.
The length of this vector should match the number of unique animal IDs, and the dates must be in the same order as the ID levels.
Alternatively, if a single value is provided, it will be applied to all IDs.}

\item{tag.durations}{Optional. A numeric vector containing the estimated battery
duration of the deployed tags (in days). The length of this vector should match the number of
unique animal IDs, and the values must be in the same order as the ID levels.
Alternatively, if a single value is provided, it will be applied to all IDs.}

\item{id.groups}{Optional. A list containing ID groups, used to
visually aggregate animals belonging to the same class (e.g. different species).}

\item{discard.missing}{If true, only individuals with detections are included.
If false, empty plots are drawn for missing individuals.}

\item{color.by}{Optional. Variable defining the color group of individual detections.
Can be used for example to depict detections of different receivers or in different habitats.}

\item{color.pal}{Color palette for the points.}

\item{date.format}{Date-time format (as used in \code{\link[base]{strptime}}),
defining the x-axis labels. Defaults to month-year ("\%b/\%y").}

\item{date.interval}{Number defining the interval between each
displayed date (x-axis label). Defaults to 4.}

\item{date.start}{Integer defining the first displayed date (can be used in combination
with 'date.interval" to better control the x-axis labels). Defaults to 1.}

\item{sunriset.coords}{A SpatialPoints or matrix object containing longitude and
latitude coordinates (in that order) at which to estimate sunrise and sunset times.}

\item{diel.lines}{Number indicating the number of diel phase lines (boundaries)
to display. Either 0 (no lines), 2 (corresponding to sunrise and sunset) or 4
(depicting dawn, sunrise, sunset, dusk). Defaults to 4.}

\item{background.color}{Color for the plot background. Defaults to "gray96".}

\item{highlight.isolated}{If a 'color.by' variable is defined, detections are ordered
and plotted according to their "density", with isolated detections being brought forward
to prevent them from being hidden behind points' clouds. Defaults to True.}

\item{pt.cex}{Expansion factor(s) for the points (detections). Defaults to 1.6.}

\item{grid}{If true, a grid is plotted (horizontal and vertical guides across the entire plot). Defaults to False.}

\item{grid.color}{Color of the grid lines. Defaults to white.}

\item{cex.title}{Determines the size of the plot title (animal ID). Defaults to 2.2.}

\item{cex.lab}{Determines the size of the axes titles. Defaults to 1.8.}

\item{cex.axis}{Determines the size of the tick mark labels on the axes. Defaults to 1.4.}

\item{cex.legend}{Determines the size of the color legend. Defaults to 1.6.}

\item{cols}{Number of columns in the final panel. Defaults to 2.}

\item{legend.cols}{Number of columns in the legend (when 'color.by' is defined).  Defaults to 3.}
}
\description{
Plots color-coded detections over time and date, independently for each individual.
}
