% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFFTs.R
\name{plotFFTs}
\alias{plotFFTs}
\title{Analyze fine-scale rhythmic patterns in detections}
\usage{
plotFFTs(
  data,
  tagging.dates = getDefaults("tagging.dates"),
  id.col = getDefaults("id"),
  timebin.col = getDefaults("timebin"),
  type = "detections",
  id.groups = NULL,
  axis.periods = c(48, 24, 12, 8, 6),
  min.days = 10,
  detrend = T,
  same.scale = T,
  background.col = "grey94",
  cex.title = 2,
  cex.lab = 1.8,
  cex.axis = 1.4,
  cols = 2
)
}
\arguments{
\item{data}{A data frame containing binned animal detections.}

\item{tagging.dates}{A POSIXct vector containing the tag/release date of each animal.
The length of this vector should match the number of unique animal IDs, and the dates must be in the same order as the ID levels.
Alternatively, if a single value is provided, it will be applied to all IDs.}

\item{id.col}{Name of the column containing animal IDs. Defaults to 'ID'.}

\item{timebin.col}{Name of the column containing time bins (in POSIXct format). Defaults to 'timebin'.}

\item{type}{Type of response used to calculate FFTs. One of  "detections" or "presences".
Presences are estimated by time bin (0 - 1), independently of the number of detections.}

\item{id.groups}{Optional. A list containing ID groups, used to
visually aggregate animals belonging to the same class (e.g. different species).}

\item{axis.periods}{Periods (in hours) to display below the plots.}

\item{min.days}{Discard individuals that were detected in less than x days.}

\item{detrend}{Detrend time series using differences (\code{\link[base]{diff}})
rather than the actual detections.}

\item{same.scale}{Forces same spectral scale (ylims) across all plots,
allowing for density comparison between individuals. If set to false, all y-axis are displayed.
Otherwise they are only displayed in the left-most plots to save space.}

\item{background.col}{A string specifying the background color of the plot. Defaults to "grey94".}

\item{cex.title}{Determines the size of the plot title (animal ID). Defaults to 2.}

\item{cex.lab}{Determines the size of the axes titles. Defaults to 1.8.}

\item{cex.axis}{Determines the size of the tick mark labels on the axes. Defaults to 1.4.}

\item{cols}{Number of columns in the final panel.}
}
\description{
Function that analyzes and plots periodic patterns in detections using a
FFT (Fast Fourier Transform) framework. By decomposing data series into the
frequency domain, FFT allows the identification of dominant spectral peaks
that may reflect, for example, tidal (6 h–12 h) or diel (24 h) cyclic patterns in habitat use.
}
