% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMaps.R
\name{plotMaps}
\alias{plotMaps}
\title{Plot maps with movement trajectories + home ranges}
\usage{
plotMaps(
  data,
  kud.densities,
  animal.tracks = NULL,
  id.groups = NULL,
  id.col = "ID",
  lon.col = "lon",
  lat.col = "lat",
  land.shape,
  background.layer = NULL,
  background.pal = NULL,
  land.color = "gray50",
  discard.missing = T,
  kud.transparency = 0,
  scale.meters = NULL,
  scale.pos = "bottomright",
  scale.inset = 0.1,
  same.scale = FALSE,
  kud.legend = TRUE,
  cols = 3
)
}
\arguments{
\item{data}{A data frame with animal positions (COAs), containing 'longitude' and 'latitude' columns.}

\item{kud.densities}{Kernel utilization areas ('estUD' class), as returned by \code{\link{calculateKUDs}}.}

\item{animal.tracks}{The output of \code{\link{calculateDistances}}.}

\item{id.groups}{Optional. A list containing ID groups, used to
visually aggregate animals belonging to the same class (e.g. different species).}

\item{id.col}{Name of the column containing animal IDs. Defaults to "ID".}

\item{lon.col}{Name of the column containing projected longitudes. Defaults to "lon".}

\item{lat.col}{Name of the column containing projected latitudes. Defaults to "lat".}

\item{land.shape}{A projected shape file containing coastlines.}

\item{background.layer}{A projected raster containing a variable to be displayed in the background
(e.g. bathymetry, temperature, etc.).}

\item{background.pal}{Color palette for the background layer.}

\item{land.color}{Color of land areas.}

\item{discard.missing}{If true, only individuals with detections are included.
If false, plots of missing individuals are drawn only with land.shape and provided background.}

\item{kud.transparency}{Transparency of the kernel utilization areas (0-1).}

\item{scale.meters}{Distance covered by the scale bar, in meters. If null, it is
automatically defined as 20\% of the plot region.}

\item{scale.pos}{Position of the scale bar, specified by keyword.}

\item{scale.inset}{Inset distance(s) from the margins as a fraction of the
plot region, relatively to the scale bar position.}

\item{same.scale}{Forces same KUD scale across all plots, allowing for
density comparison between individuals. If set to false, all y-axis are displayed.
Otherwise they are only displayed in the left-most plots to save space.}

\item{kud.legend}{Logical. If TRUE, includes a legend for the KUD densities. Defaults to TRUE.}

\item{cols}{Number of columns in the panel (passed to the mfrow argument). Defaults to 3.}
}
\description{
Produces maps with inferred movement trajectories
(shortest paths in water) and estimated kernel home-ranges.
It can also overlay information such as bathymetry or
sea surface temperature, by plotting a raster in the background.
}
