% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMigrations.R
\name{plotMigrations}
\alias{plotMigrations}
\title{Plot map representing animal transitions/movements between receivers/sites.
If tagging information is supplied, the number of tagged individuals is indicated for
each location/site and the release site is included in the calculated transitions.}
\usage{
plotMigrations(
  data,
  id.groups = NULL,
  tag.info = NULL,
  aggregate.by,
  land.shape = NULL,
  land.color = "gray50",
  background.layer = NULL,
  background.pal = "#F3F7F7",
  color.nodes.by = "detection",
  nodes.color = c("darkblue", "black"),
  nodes.alpha = 0.8,
  nodes.size = c(0.035, 0.08),
  repel.nodes = F,
  repel.buffer = 1.1,
  scale.km = NULL,
  scale.pos = "bottomright",
  scale.inset = 0.05,
  scale.height = 1.5,
  id.col = "ID",
  lon.col = "lon",
  lat.col = "lat",
  bbox.ext = 0.2,
  ...
)
}
\arguments{
\item{data}{A data frame containing binned animal detections.}

\item{id.groups}{Optional. A list containing ID groups (e.g. different species or size-classes).
If supplied, a separate map will be plotted for each group.}

\item{aggregate.by}{Name of the variable/column containing the spatial information used to calculate transitions.
It can include receiver IDs in case all movements/transitions are of interest,
or contain (for example) location/habitat classes for broader-scale analyses.}

\item{land.shape}{Optional. A shape file containing coastlines.}

\item{land.color}{Optional. Color of land areas.}

\item{background.layer}{A projected raster containing a variable to be displayed in the background
(e.g. bathymetry, temperature, etc.).}

\item{background.pal}{Color palette for the background layer. If a background layer is not supplied,
the first color is used as background.}

\item{color.nodes.by}{Variable used to color nodes (locations). Alternatively, it can be
set to 'detectios' (default) to distinguish between nodes with detections and nodes without detections or
to 'group' to assign different colors to different ID groups (one color per map).}

\item{nodes.color}{Color of the nodes where animals have been detected.}

\item{nodes.alpha}{Opacity of the nodes: 0=fully transparent, 1=fully opaque.}

\item{nodes.size}{A numeric vector of length 2. Represents the desired min and max vertex sizes relative
to the x-axis in terms of percentage (see details). See \code{\link[netdiffuseR]{rescale_vertex_igraph}}.}

\item{repel.nodes}{Used to automatically separate nodes if they overlap or are too clustered in space.
Defaults to False.}

\item{repel.buffer}{Controls the amount of space between nodes if repel.nodes is set to True.}

\item{scale.km}{Distance covered by the scale bar, in kilometers. If null, it is
automatically defined as 20\% of the plot region.}

\item{scale.pos}{Position of the map scale, specified by keyword.
See \code{\link[base]{xy.coords}}.}

\item{scale.inset}{Inset distance(s) of the map scale from the margins
as a fraction of the plot region.}

\item{scale.height}{Controls the thickness of the scale bar. Defaults to 1.5.}

\item{id.col}{Name of the column containing animal IDs Defaults to 'ID'.}

\item{lon.col}{Name of the column containing longitude values. Defaults to 'lon'.}

\item{lat.col}{Name of the column containing latitude values. Defaults to 'lat'.}

\item{bbox.ext}{Amount of extension of the bounding box around the coordinates.
It can be numeric (same extension into all four directions), vector of two (first x, then y directional extension)
or vector of four (xmin, xmax, ymin, ymax extension). Default is .2 (extends the bounding box by 20\%).
Only considered if a background layer is not supplied.}

\item{...}{Arguments passed to the \code{\link[igraph]{plot.igraph}} function.}

\item{tags.info}{Optional. A data frame containing information on the tagging location of each animal.
Used to xxx.}
}
\description{
Plot map representing animal transitions/movements between receivers/sites.
If tagging information is supplied, the number of tagged individuals is indicated for
each location/site and the release site is included in the calculated transitions.
}
