% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOverlapNetwork.R, R/plotOverlaps.R
\name{plotOverlapNetwork}
\alias{plotOverlapNetwork}
\title{Network representation of pairwise overlaps.}
\usage{
plotOverlapNetwork(
  overlaps,
  random.results = NULL,
  min.val = NULL,
  cut.val = NULL,
  group.order = NULL,
  title.cex = 1,
  color.nodes.by = "detection",
  network.layout = "spring",
  nodes.color = NULL,
  nodes.alpha = 0.8,
  nodes.size = c(0.7, 1.4),
  nodes.scale.by = NULL,
  nodes.label.wrap = FALSE,
  nodes.label.cex = 0.5,
  nodes.label.color = "white",
  repel.nodes = FALSE,
  repel.buffer = 1.1,
  edge.color = "darkblue",
  edge.curved = 0.5,
  edge.width = c(0.4, 3.5),
  edge.arrow.size = 0.5,
  edge.arrow.width = 1.5,
  edge.label.cex = 0.6,
  edge.label.color = "black",
  edge.label.font = 1,
  ...
)

plotOverlapNetwork(
  overlaps,
  random.results = NULL,
  min.val = NULL,
  cut.val = NULL,
  group.order = NULL,
  title.cex = 1,
  color.nodes.by = "detection",
  network.layout = "spring",
  nodes.color = NULL,
  nodes.alpha = 0.8,
  nodes.size = c(0.7, 1.4),
  nodes.scale.by = NULL,
  nodes.label.wrap = FALSE,
  nodes.label.cex = 0.5,
  nodes.label.color = "white",
  repel.nodes = FALSE,
  repel.buffer = 1.1,
  edge.color = "darkblue",
  edge.curved = 0.5,
  edge.width = c(0.4, 3.5),
  edge.arrow.size = 0.5,
  edge.arrow.width = 1.5,
  edge.label.cex = 0.6,
  edge.label.color = "black",
  edge.label.font = 1,
  ...
)
}
\arguments{
\item{overlaps}{Similarity matrix containing pairwise overlaps, as returned by \code{\link{calculateOverlap}}.}

\item{min.val}{Edges with overlaps lower than this value are not shown.}

\item{cut.val}{Edges with overlaps larger than this value will have the strongest color intensity
and become wider the stronger they are, and edges with overlaps lower than this value
will have the smallest width and become vaguer the weaker the weight.}

\item{title.cex}{Size of the plot title. Defaults to 1.}

\item{color.nodes.by}{Variable used to color nodes (locations). It can be set to any variable in the dataset.
Alternatively, it can be set to 'detections' (default) to distinguish between nodes with and without detections,
or to 'group' to assign different colors to different ID groups (one color per map). Defaults to "detection".}

\item{nodes.color}{Color(s) for the nodes. Defaults to c("darkblue", "black").}

\item{nodes.alpha}{Numeric. Transparency level for the nodes (0 to 1): 0=fully transparent, 1=fully opaque. Defaults to 0.8.}

\item{nodes.size}{A numeric vector of length 2. Represents the desired min and max vertex sizes relative
to the x-axis in terms of percentage (see details). See \code{\link[netdiffuseR]{rescale_vertex_igraph}}. Defaults to c(0.04, 0.08).}

\item{nodes.label.wrap}{Logical. If TRUE, splits node labels into multiple lines (to better fit within the nodes). Defaults to FALSE.}

\item{nodes.label.cex}{The font size for vertex labels. Defaults to 0.5.}

\item{nodes.label.color}{The font color for vertex labels. Defaults to white.}

\item{repel.nodes}{Logical. If TRUE, nodes are plotted using a repulsion algorithm to avoid overlap. Defaults to FALSE.}

\item{repel.buffer}{Controls the amount of space between nodes if repel.nodes is set to TRUE. Defaults to 1.1.}

\item{edge.color}{Color(s) for the edges. Defaults to "darkblue".}

\item{edge.curved}{Specifies whether to draw curved edges, or not. This can be a logical or a numeric vector or scalar.
A numeric value specifies the curvature of the edge; zero curvature means straight edges, negative values mean the edge bends
clockwise, positive values the opposite. TRUE means curvature 0.5, FALSE means curvature zero. Defaults to 0.5.}

\item{edge.width}{A numeric vector of length 2. Represents the desired min and max
width/thickness of the edges. Defaults to c(0.4, 3.5).}

\item{edge.arrow.size}{The size of the arrows. Defaults to 0.5.}

\item{edge.arrow.width}{The width of the arrows. Defaults to 1.5.}

\item{edge.label.cex}{The font size for edge labels. Defaults to 0.6.}

\item{edge.label.color}{The color of the edge labels. Defaults to black.}

\item{edge.label.font}{The font for the edge labels. It is interpreted the same way
as the font graphical parameter: 1 is plain text, 2 is bold face, 3 is italic,
4 is bold and italic and 5 specifies the symbol font. Defaults to 1.}

\item{...}{Further arguments passed to \code{\link[qgraph]{qgraph}}.}

\item{id.metadata}{A data frame containing animal IDs as well as an optional 'color.by' and 'size.by' variables.}

\item{color.by}{Optional. Factor variable in id.metadata defining the color of each node (e.g. "species", "sex", etc).}

\item{scale.by}{Optional. Numeric variable in id.metadata defining the size of each node (e.g.  "length").}

\item{model}{Type of regression in case animal.sizes have been supplied ("linear" or "quadratic").
A vector with two elements should be supplied, the first to be used in the "average overlap ~ size" graph,
and the second in the "pairwise overlap ~ size difference" graph.}

\item{color.pal}{Color of the nodes.}

\item{min.value}{Defaults to the 50\% quantile}

\item{cut.value}{Defaults to the 90\% quantile}

\item{edge.type}{A character string indicating the metric to be used to calculate the network edges.
It can be either "movements" to represent the number of movements between sites or "individuals"
to represent the number of individual animals moving between sites. Defaults to "movements".}
}
\description{
Function to plot a network of estimated pairwise overlaps.
Each node represents a single individual and can be dimensioned proportionally
to its size. Edges represent the extent of spatial overlap between each pair
(i.e. greater thickness indicates a higher overlap). If animals sizes are supplied,
two additional plots are produced, one representing the relationship between
animals' size/length and average overlap and the other representing the
relationship between pairwise differences in size and overlap scores.

Function to plot a network of estimated pairwise overlaps.
Each node represents a single individual and can be dimensioned proportionally
to its size. Edges represent the extent of spatial overlap between each pair
(i.e. greater thickness indicates a higher overlap). If animals sizes are supplied,
two additional plots are produced, one representing the relationship between
animals' size/length and average overlap and the other representing the
relationship between pairwise differences in size and overlap scores.
}
\seealso{
\code{\link{calculateOverlap}}

\code{\link{plotOverlapFrequences}}

\code{\link{calculateOverlap}}

\code{\link{randomizeOverlaps}}
}
