% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOverlapTable.R
\name{plotOverlapTable}
\alias{plotOverlapTable}
\title{Plots pairwise stats in a contingency table}
\usage{
plotOverlapTable(
  randomized.overlaps,
  id.metadata = NULL,
  id.groups = NULL,
  type = "significance",
  group.comparisons = "both",
  color.pal = NULL,
  discard.empty = TRUE,
  id.col = getDefaults("ID"),
  full.scale = FALSE,
  sort.by = NULL
)
}
\arguments{
\item{randomized.overlaps}{Similarity matrix containing pairwise overlaps, as returned by \code{\link{calculateOverlap}}.}

\item{id.metadata}{Optional. A data frame containing individual's metadata. It should contain
at least animal IDs  and a numeric variable (e.g. length), defined by}

\item{id.groups}{Optional. A list containing ID groups, used to calculate stats independently
within each group and/or to compare relationships between ids of different groups.}

\item{type}{Type of metric to be plotted, one of "significance" or "mean overlap".
Defaults to "significance".}

\item{group.comparisons}{Controls the type of comparisons to be run, when id.groups are defined.
One of "within", "between" or "both". Useful to visualize comparisons between individuals
belonging to the same id.group or to only compare individuals in different id.groups
(e.g. intraspecific vs interspecific). Defaults to "both".}

\item{color.pal}{Color palette. If type is set to significance, it should contain 3 colors
("+", "-", "ns").}

\item{discard.empty}{Boolean. Discard IDs with missing interactions.}

\item{id.col}{Name of the column in id.metadata containing animal IDs. Defaults to 'ID'.}

\item{full.scale}{Boolean. If type is set to 'mean overlap', sets the overlap scale from 0 to 100\%.}

\item{sort.by}{Optional. Sort IDs based on this variable (e.g., length).}
}
\description{
Plots a contingency table illustrating the pairwise distribution of overlap
scores/significance across individuals. IDs can be sorted by length or any other quantitative metric.
}
\seealso{
\code{\link{calculateOverlap}} and \code{\link{randomizeOverlaps}}
}
