% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStationStats.R
\name{plotStationStats}
\alias{plotStationStats}
\title{#' Calculate and plot receiver-based statistics}
\usage{
plotStationStats(
  data,
  type = c("detections", "co-occurrences"),
  id.groups = NULL,
  group.comparisons = "all",
  id.col = getDefaults("ID"),
  timebin.col = getDefaults("timebin"),
  station.col = getDefaults("station"),
  aggregate.by = NULL,
  color.pal = NULL,
  number.locations = FALSE,
  rotate.names = FALSE,
  show.percentage = TRUE,
  show.counts = TRUE,
  annot.threshold = NULL,
  annot.cex = 0.7,
  title.pos = "top",
  title.inset = c(0, 0.02),
  title.cex = 1.1,
  label.cex = 1,
  axis.cex = 0.9,
  legend.pos = "topright",
  legend.style = "vertical",
  legend.cex = 0.7,
  cols = 1
)
}
\arguments{
\item{data}{A data frame containing animal detections with corresponding time-bins.}

\item{type}{Type of statistic to calculate and plot. If multiple types are supplied, a grouped bar plot
is generated. Possible values are:
\itemize{
\item \code{"detections"}: Total number of detections at each station.
\item \code{"average detections"}: For each individual, the proportion of detections at each station
is calculated relative to the total detections for that individual across all stations. These proportions
are then averaged across all individuals to produce a station-specific value. This metric provides insight
into the average relative importance of each station across all individuals.
\item \code{"individuals"}: Number of unique individuals detected at each station.
\item \code{"co-occurrences"}: Number of co-occurrence events at each station, where two or more individuals
are detected within the same time-bin.
}}

\item{id.groups}{Optional. A list containing ID groups, used to generate different
plots to each class (e.g. different species).}

\item{group.comparisons}{Controls the type of comparisons to be run, when id.groups are defined.
Only relevant to estimate co-occurrences. One of "within" (intra-group), "between" (inter-group) or "all" (both). Defaults to "all".}

\item{id.col}{Name of the column containing animal IDs. Defaults to 'ID'.}

\item{timebin.col}{Name of the column containing time bins (in POSIXct format). Defaults to 'timebin'.}

\item{station.col}{Name of the column containing station/receiver IDs. Defaults to 'station'.}

\item{aggregate.by}{Column name for aggregating statistics (defaults to \code{station.col}).}

\item{color.pal}{A color palette for the bars; the length should match the number of \code{type} values provided.}

\item{number.locations}{Logical. If \code{TRUE}, replaces location names with numbers (for clearer display when names are long).}

\item{rotate.names}{Logical. If \code{TRUE}, rotates location names on the x-axis.}

\item{show.percentage}{Logical. If \code{TRUE}, percentages are displayed above each bar.}

\item{show.counts}{Logical. If \code{TRUE}, shows the count above each bar.}

\item{annot.threshold}{Numeric (optional). Minimum count required to display
labels on bars (e.g., \code{1} only labels bars with counts equal to or greater than 1).
Useful for reducing visual clutter in plots with many bars.}

\item{annot.cex}{Numeric. Font size for bar annotation text. Defaults to 0.7.}

\item{title.pos}{Position of the title on the plot (by keyword). Defaults to "top".}

\item{title.inset}{Inset distance of the title from the specified position, as a fraction of the
plot region. Defaults to c(0, 0.02)}

\item{title.cex}{Numeric. Font size for the plot title. Defaults to 1.1.}

\item{label.cex}{Numeric. Font size for axis labels. Defaults to 1.}

\item{axis.cex}{Numeric. Font size for axis text. Defaults to 0.9.}

\item{legend.pos}{Legend position.}

\item{legend.style}{Orientation of the legend, either "horizontal" or "vertical".}

\item{legend.cex}{Numeric. Font size for legend text. Defaults to 0.7.}

\item{cols}{Number of columns in the final panel (passed to the mfrow argument).}
}
\description{
This function calculates and visualizes receiver-based statistics from passive acoustic telemetry data.
It generates grouped bar plots to compare specified statistics across stations or other spatial categories.
Users can analyze metrics such as total detections, average detection proportions, the number of unique
individuals detected, and co-occurrence events. The function supports grouping by animal categories
(e.g., species or sexes) and offers extensive customization options for plot aesthetics.
}
