% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomizeOverlaps.R
\name{randomizeOverlaps}
\alias{randomizeOverlaps}
\title{Test significance of estimated overlaps through null model randomization tests}
\usage{
randomizeOverlaps(
  table,
  overlaps,
  id.metadata,
  contraint.by = NULL,
  id.groups = NULL,
  id.col = "ID",
  tagdate.col = "tagging_date",
  group.comparisons = "all",
  iterations = 100,
  conf.level = 0.95,
  cores = 1,
  random.seed = NULL
)
}
\arguments{
\item{table}{A data frame object containing binned detections in the wide format
(time bin x individual matrix, with values corresponding to the receiver with the
highest number of detections), as returned by \code{\link{createWideTable}}.}

\item{overlaps}{Similarity matrix containing pairwise overlaps, as returned by \code{\link{calculateOverlap}}.}

\item{id.metadata}{A data frame containing individual's metadata. It should contain
at least  animal IDs  and a tagging/release date column in POSIXct format.}

\item{contraint.by}{Variable or variables used to constraint permutations, e.g.
to account for potential diel or seasonal trends in animal occurrences. If supplied,
permutations across time bins will be restricted to the same combination of
levels of these variables.}

\item{id.groups}{Optional. A list containing ID groups, used to calculate stats independently
within each group, as well as comparing relationships between ids of different groups.}

\item{id.col}{Name of the column in id.metadata containing animal IDs. Defaults to 'ID'.}

\item{tagdate.col}{Name of the column in id.metadata containing animal tagging/release dates in
POSIXct format. Defaults to 'tagging_date'.}

\item{group.comparisons}{Controls the type of comparisons to be run, when id.groups are defined.
One of "within", "between" or "all". Useful to discard comparisons between individuals
belonging to the same group or skip comparisons between different groups, when these are
not required (less computing time). Defaults to "all".}

\item{iterations}{Number of Monte Carlo iterations (simulated datasets).}

\item{conf.level}{Confidence level. Defaults to 95\%.}

\item{cores}{Number of CPU cores for parallel computing (increase processing power). Defaults to 1.}

\item{random.seed}{Set the seed for a reproducible randomization.}
}
\description{
Tests the null hypothesis of temporally independent space use
(i.e. each animal occurs independently of the other) using Monte Carlo permutation tests.
The algorithm permutes entries within each column, so that the total number of detections
of each individual and the relative occurrence frequencies across receivers are kept unchanged.
}
