% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadeSeasons.R
\name{shadeSeasons}
\alias{shadeSeasons}
\title{Return season boundaries for background shading.}
\usage{
shadeSeasons(
  start.time,
  end.time,
  interval,
  color.pal = c("white", "grey96", "grey83", "grey90"),
  hemisphere = "Northern"
)
}
\arguments{
\item{start.time}{A POSIXct object containing the earliest date used in the plot region.}

\item{end.time}{A POSIXct object containing the latest date used in the plot region.}

\item{interval}{Time-bins interval (in minutes).}

\item{color.pal}{Vector of 4 colors, one for each season
(in the following order: winter, spring, summer and autumn).}

\item{hemisphere}{Earth hemisphere for which to calculate seasons.}
}
\description{
Returns a table containing season boundaries/limits, used to add background polygons
in \code{\link{plotAbacus}}.
}
\seealso{
\code{\link{getSeason}}
}
