% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryTable.R
\name{summaryTable}
\alias{summaryTable}
\title{Create summary table}
\usage{
summaryTable(
  data,
  tagging.dates,
  tags.info = NULL,
  residency.by = NULL,
  id.groups = NULL,
  error.stat = "sd"
)
}
\arguments{
\item{data}{A data frame containing binned animal detections.}

\item{tagging.dates}{A POSIXct vector containing the tag/release date of each animal.}

\item{tags.info}{A data frame containing additional informations about the tagged animals,
such as length, sex or transmitter type. If more than one row exists per animal,
variables are collapsed before being merged with the remaining stats.}

\item{residency.by}{Variable used to calculate partial residencies (e.g. array or habitat).
Defaults to NULL.}

\item{id.groups}{Optional. A list containing ID groups, used to
visually aggregate animals belonging to the same class (e.g. different species or life stages).}

\item{error.stat}{Statistic to present with the mean. Either standard deviation (sd) or standard error (se).}

\item{id.col}{Name of the column containing animal IDs. Defaults to 'ID'.}
}
\description{
Creates a table containing information about the
tagged animals (tagging dates, last detections, etc.) as well
as residency metrics.
}
