% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryTable.R
\name{summaryTable}
\alias{summaryTable}
\title{Generate summary table for tagged animals}
\usage{
summaryTable(
  data,
  tagging.dates = getDefaults("tagdates"),
  id.metadata = NULL,
  id.col = getDefaults("id"),
  datetime.col = getDefaults("datetime"),
  station.col = getDefaults("station"),
  residency.by = NULL,
  id.groups = NULL,
  error.stat = "sd"
)
}
\arguments{
\item{data}{A data frame containing animal detections. Each row should represent an individual detection event,
unless a 'detections' column is included to indicate the number of detections for each row.}

\item{tagging.dates}{A POSIXct vector containing the tag/release date of each animal.
The length of this vector should match the number of unique animal IDs.
Alternatively, if a single value is provided, it will be applied to all IDs.}

\item{id.metadata}{A data frame containing metadata about the tagged animals, such as their length,
sex, or transmitter type. If there are multiple rows per animal, variables will be collapsed before merging with other statistics.}

\item{id.col}{Name of the column containing animal IDs. Defaults to 'ID'.}

\item{datetime.col}{Name of the column containing datetimes in POSIXct format. Defaults to 'datetime'.}

\item{station.col}{Name of the column containing station/receiver IDs. Defaults to 'station'.}

\item{residency.by}{Optional. Variable used to calculate partial residencies (e.g. array or habitat).
Defaults to NULL.}

\item{id.groups}{Optional. A list where each element is a group of IDs, used for visually aggregating
animals belonging to the same class (e.g., different species or life stages). If supplied, averages
will be calculated independently for each group.}

\item{error.stat}{The statistic to use for variability/error calculation, either 'sd' (standard deviation)
or 'se' (standard error). Defaults to 'sd'.}
}
\description{
This function generates a summary table with information about tagged animals,
including tagging dates, last detection dates, and various monitoring and residency metrics. It allows for the
inclusion of additional metadata and the calculation of overall means and error metrics.
}
