% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryTable.R
\name{summaryTable}
\alias{summaryTable}
\title{Generate summary table for tagged animals}
\usage{
summaryTable(
  data,
  tagging.dates = getDefaults("tagdates"),
  tag.durations = NULL,
  id.metadata = NULL,
  id.col = getDefaults("ID"),
  datetime.col = getDefaults("datetime"),
  station.col = getDefaults("station"),
  sensor.cols = NULL,
  sensor.titles = NULL,
  residency.index = c("IR1", "IR2", "IWR"),
  start.point = "release",
  residency.by = NULL,
  id.groups = NULL,
  error.stat = "sd"
)
}
\arguments{
\item{data}{A data frame containing animal detections. Each row should represent an individual detection event,
unless a 'detections' column is included to indicate the number of detections for each row.}

\item{tagging.dates}{A POSIXct vector containing the tag/release date of each animal.
The length of this vector should match the number of unique animal IDs, and the dates must be in the same order as the ID levels.
Alternatively, if a single value is provided, it will be applied to all IDs.}

\item{tag.durations}{Optional. A numeric vector containing the estimated battery
duration of the deployed tags (in days). The length of this vector should match the number of
unique animal IDs, and the values must be in the same order as the ID levels.
Alternatively, if a single value is provided, it will be applied to all IDs.}

\item{id.metadata}{A data frame containing metadata about the tagged animals, such as their length,
sex, or transmitter type. If there are multiple rows per animal, variables will be collapsed before merging with other statistics.}

\item{id.col}{Name of the column containing animal IDs. Defaults to 'ID'.}

\item{datetime.col}{Name of the column containing datetimes in POSIXct format. Defaults to 'datetime'.}

\item{station.col}{Name of the column containing station/receiver IDs. Defaults to 'station'.}

\item{sensor.cols}{Optional. A character vector specifying column names in \code{data} that contain
additional sensor readings (e.g., depth, temperature). For each column specified, the mean,
minimum, and maximum values will be calculated and incorporated into the summary.}

\item{sensor.titles}{Optional. Titles for the sensor readings in \code{sensor.cols}.
If not provided, defaults to the column names.}

\item{residency.index}{A character string specifying the type of residency index to calculate.
Options include:
\itemize{
\item "IR1": Residency Index 1, calculated as the number of days the animal was detected (Dd) divided by the detection interval (Di), i.e., the number of days between release/first detection and last detection (days at liberty).
This represents a maximum residency value, considering only the period for which the animal was known to be alive and the tag operational.
\item "IR2": Residency Index 2, calculated as the number of days the animal was detected (Dd) divided by the study interval (Dt), i.e., the total number of days between release/first detection and last data download or tag expiration date.
This approach provides a minimum residency value, assuming the animal was alive and detectable throughout the study period.
\item "IWR": Weighted Residency Index, which corresponds to the IR2 index weighted by the ratio between the detection interval (Di, the number of days between the first and last detection) and the study interval (Dt, the total monitoring period).
This accounts for the number of days detected and the spread of detections within the monitoring period, providing a measure of residency that balances the frequency of detections with their temporal distribution.
}

The choice of index can affect the interpretation of residency patterns, so it's important to select the one(s) that best fits the study objectives.
Further information on residency estimation methods can be found in Kraft et al. (2023) and Appert et al. (2023) - see below in the references section.}

\item{start.point}{A character string specifying the starting point for calculating days at liberty.
Options include:
\itemize{
\item \code{"release"}: The release day is used as the starting point for calculating days at liberty.
\item \code{"first.detection"}: The first detection after tagging is used as the starting point.
}

Defaults to \code{"release"}.}

\item{residency.by}{Optional. Variable used to calculate partial residencies (e.g. array or habitat).
Defaults to NULL.}

\item{id.groups}{Optional. A list where each element is a group of IDs, used for visually aggregating
animals belonging to the same class (e.g., different species or life stages). If supplied, averages
will be calculated independently for each group.}

\item{error.stat}{The statistic to use for variability/error calculation, either 'sd' (standard deviation)
or 'se' (standard error). Defaults to 'sd'.}
}
\value{
A data frame summarizing information on tagged animals, with the following columns:
\itemize{
\item \code{ID}: Unique identifier for each tagged animal.
\item Any additional metadata columns from \code{id.metadata} if provided.
\item \verb{Tagging date}: The date when the animal was tagged.
\item \verb{Tag duration (d)}: Duration of the tag in days (if provided).
\item \verb{Last detection}: The date of the last detection.
\item \verb{N Detect}: Total number of detections for the animal.
\item \verb{N Receiv}: Number of unique receivers that detected the animal.
\item \verb{Monitoring duration (d)}: Total duration of monitoring in days.
\item \verb{Detection span (d)}: Number of days between release/first detection and last detection (days at liberty)
\item \verb{N days detected}: Total number of days the animal was detected.
\item Additional columns for each residency index specified in the \code{residency.index} parameter.
\item If \code{residency.by} is specified, additional columns for partial residency metrics will be included.
\item Sensor data metrics: For each column in \code{sensor.cols}, the mean, minimum, and maximum values, using titles specified in \code{sensor.titles} (if provided).
}
}
\description{
This function generates a comprehensive summary table for tagged animals,
including tagging and last detection dates, along with various monitoring and residency metrics.
It also allows the inclusion of additional sensor data (e.g., depth, temperature) and provides
an option to summarize these readings by calculating the mean, minimum, and maximum values for each specified sensor column.
Additionally, the function can incorporate additional metadata and automatically
computes overall means and error metrics.
}
\references{
Kraft, S., Gandra, M., Lennox, R. J., Mourier, J., Winkler, A. C., & Abecasis, D. (2023).
Residency and space use estimation methods based on passive acoustic telemetry data.
Movement Ecology, 11(1), 12.
https://doi.org/10.1186/s40462-023-00349-y

Appert, C., Udyawer, V., Simpfendorfer, C. A., Heupel, M. R., Scott, M., Currey-Randall, L. M., ... & Chin, A. (2023).
Use, misuse, and ambiguity of indices of residence in acoustic telemetry studies.
Marine Ecology Progress Series, 714, 27-44.
https://doi.org/10.3354/meps14300
}
